Network.AWS.SES

Description

Amazon Simple Email Service (Amazon SES) is a cost-effective outbound-only email-sending service built on the reliable and scalable infrastructure that Amazon.com has developed to serve its own customer base. With Amazon SES, you can send transactional email, marketing messages, or any other type of high-quality content and you only pay for what you use. Along with high deliverability, Amazon SES provides easy, real-time access to your sending statistics and built-in notifications for bounces, complaints, and deliveries to help you fine-tune your email-sending strategy.

Documentation

module Network.AWS.SES.DeleteIdentity

module Network.AWS.SES.DeleteVerifiedEmailAddress

module Network.AWS.SES.GetIdentityDkimAttributes

module Network.AWS.SES.GetIdentityNotificationAttributes

module Network.AWS.SES.GetIdentityVerificationAttributes

module Network.AWS.SES.GetSendQuota

module Network.AWS.SES.GetSendStatistics

module Network.AWS.SES.ListIdentities

module Network.AWS.SES.ListVerifiedEmailAddresses

module Network.AWS.SES.SendEmail

module Network.AWS.SES.SendRawEmail

module Network.AWS.SES.SetIdentityDkimEnabled

module Network.AWS.SES.SetIdentityFeedbackForwardingEnabled

module Network.AWS.SES.SetIdentityNotificationTopic

module Network.AWS.SES.Types

module Network.AWS.SES.VerifyDomainDkim

module Network.AWS.SES.VerifyDomainIdentity

module Network.AWS.SES.VerifyEmailAddress

module Network.AWS.SES.VerifyEmailIdentity

module Network.AWS.SES.Waiters