Network.AWS.SES.CreateConfigurationSetEventDestination

Description

Creates a configuration set event destination.

An event destination is the AWS service to which Amazon SES publishes the email sending events associated with a configuration set. For information about using configuration sets, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

createConfigurationSetEventDestination

Arguments

:: Text
ccsedConfigurationSetName

-> EventDestination
ccsedEventDestination

-> CreateConfigurationSetEventDestination

Creates a value of CreateConfigurationSetEventDestination with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsedConfigurationSetName - The name of the configuration set to which to apply the event destination.

ccsedEventDestination - An object that describes the AWS service to which Amazon SES will publish the email sending events associated with the specified configuration set.

data CreateConfigurationSetEventDestination

Represents a request to create a configuration set event destination. A configuration set event destination, which can be either Amazon CloudWatch or Amazon Kinesis Firehose, describes an AWS service in which Amazon SES publishes the email sending events associated with a configuration set. For information about using configuration sets, see the Amazon SES Developer Guide .

See: createConfigurationSetEventDestination smart constructor.

Request Lenses

ccsedConfigurationSetName :: Lens' CreateConfigurationSetEventDestination Text

The name of the configuration set to which to apply the event destination.

ccsedEventDestination :: Lens' CreateConfigurationSetEventDestination EventDestination

An object that describes the AWS service to which Amazon SES will publish the email sending events associated with the specified configuration set.

Destructuring the Response

createConfigurationSetEventDestinationResponse

Arguments

:: Int
ccsedrsResponseStatus

-> CreateConfigurationSetEventDestinationResponse

Creates a value of CreateConfigurationSetEventDestinationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsedrsResponseStatus - -- | The response status code.

data CreateConfigurationSetEventDestinationResponse

An empty element returned on a successful request.

See: createConfigurationSetEventDestinationResponse smart constructor.

Response Lenses

ccsedrsResponseStatus :: Lens' CreateConfigurationSetEventDestinationResponse Int

- | The response status code.