Network.AWS.SES.CreateReceiptRule

Description

Creates a receipt rule.

For information about setting up receipt rules, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

createReceiptRule

Arguments

:: Text
crrRuleSetName

-> ReceiptRule
crrRule

-> CreateReceiptRule

Creates a value of CreateReceiptRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrAfter - The name of an existing rule after which the new rule will be placed. If this parameter is null, the new rule will be inserted at the beginning of the rule list.

crrRuleSetName - The name of the rule set to which to add the rule.

crrRule - A data structure that contains the specified rule's name, actions, recipients, domains, enabled status, scan status, and TLS policy.

data CreateReceiptRule

Represents a request to create a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: createReceiptRule smart constructor.

Request Lenses

crrAfter :: Lens' CreateReceiptRule (Maybe Text)

The name of an existing rule after which the new rule will be placed. If this parameter is null, the new rule will be inserted at the beginning of the rule list.

crrRuleSetName :: Lens' CreateReceiptRule Text

The name of the rule set to which to add the rule.

crrRule :: Lens' CreateReceiptRule ReceiptRule

A data structure that contains the specified rule's name, actions, recipients, domains, enabled status, scan status, and TLS policy.

Destructuring the Response

createReceiptRuleResponse

Arguments

:: Int
crrrsResponseStatus

-> CreateReceiptRuleResponse

Creates a value of CreateReceiptRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrrsResponseStatus - -- | The response status code.

data CreateReceiptRuleResponse

An empty element returned on a successful request.

See: createReceiptRuleResponse smart constructor.

Response Lenses

crrrsResponseStatus :: Lens' CreateReceiptRuleResponse Int

- | The response status code.