Network.AWS.SES.CreateReceiptRuleSet

Description

Creates an empty receipt rule set.

For information about setting up receipt rule sets, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

createReceiptRuleSet

Arguments

:: Text
crrsRuleSetName

-> CreateReceiptRuleSet

Creates a value of CreateReceiptRuleSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrsRuleSetName - The name of the rule set to create. The name must: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-). * Start and end with a letter or number. * Contain less than 64 characters.

data CreateReceiptRuleSet

Represents a request to create an empty receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: createReceiptRuleSet smart constructor.

Request Lenses

crrsRuleSetName :: Lens' CreateReceiptRuleSet Text

The name of the rule set to create. The name must: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-). * Start and end with a letter or number. * Contain less than 64 characters.

Destructuring the Response

createReceiptRuleSetResponse

Arguments

:: Int
crrsrsResponseStatus

-> CreateReceiptRuleSetResponse

Creates a value of CreateReceiptRuleSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrsrsResponseStatus - -- | The response status code.

data CreateReceiptRuleSetResponse

An empty element returned on a successful request.

See: createReceiptRuleSetResponse smart constructor.

Response Lenses

crrsrsResponseStatus :: Lens' CreateReceiptRuleSetResponse Int

- | The response status code.