Network.AWS.SES.DeleteIdentity

Description

Deletes the specified identity (an email address or a domain) from the list of verified identities.

This action is throttled at one request per second.

Creating a Request

deleteIdentity

Arguments

:: Text
diIdentity

-> DeleteIdentity

Creates a value of DeleteIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diIdentity - The identity to be removed from the list of identities for the AWS Account.

data DeleteIdentity

Represents a request to delete one of your Amazon SES identities (an email address or domain).

See: deleteIdentity smart constructor.

Request Lenses

diIdentity :: Lens' DeleteIdentity Text

The identity to be removed from the list of identities for the AWS Account.

Destructuring the Response

deleteIdentityResponse

Arguments

:: Int
dirsResponseStatus

-> DeleteIdentityResponse

Creates a value of DeleteIdentityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dirsResponseStatus - -- | The response status code.

data DeleteIdentityResponse

An empty element returned on a successful request.

See: deleteIdentityResponse smart constructor.

Response Lenses

dirsResponseStatus :: Lens' DeleteIdentityResponse Int

- | The response status code.