Network.AWS.SES.DeleteReceiptRule

Description

Deletes the specified receipt rule.

For information about managing receipt rules, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

deleteReceiptRule

Arguments

:: Text
delRuleSetName

-> Text
delRuleName

-> DeleteReceiptRule

Creates a value of DeleteReceiptRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delRuleSetName - The name of the receipt rule set that contains the receipt rule to delete.

delRuleName - The name of the receipt rule to delete.

data DeleteReceiptRule

Represents a request to delete a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: deleteReceiptRule smart constructor.

Request Lenses

delRuleSetName :: Lens' DeleteReceiptRule Text

The name of the receipt rule set that contains the receipt rule to delete.

delRuleName :: Lens' DeleteReceiptRule Text

The name of the receipt rule to delete.

Destructuring the Response

deleteReceiptRuleResponse

Arguments

:: Int
delrsResponseStatus

-> DeleteReceiptRuleResponse

Creates a value of DeleteReceiptRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delrsResponseStatus - -- | The response status code.

data DeleteReceiptRuleResponse

An empty element returned on a successful request.

See: deleteReceiptRuleResponse smart constructor.

Response Lenses

delrsResponseStatus :: Lens' DeleteReceiptRuleResponse Int

- | The response status code.