Network.AWS.SES.DeleteVerifiedEmailAddress

Description

Deletes the specified email address from the list of verified addresses. The DeleteVerifiedEmailAddress action is deprecated as of the May 15, 2012 release of Domain Verification. The DeleteIdentity action is now preferred. This action is throttled at one request per second. POST HTTP1.1 Date: Thu, 18 Aug 2011 22:20:50 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=Rxzyd6cQe/YjkV4yoQAZ243OzzNjFgrsclizTKwRIRc=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 142 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=DeleteVerifiedEmailAddress &EmailAddress=user%40example.com &Timestamp=2011-08-18T22%3A20%3A50.000Z 5634af08-c865-11e0-8986-3f99a698f914.

Request

data DeleteVerifiedEmailAddress

Represents a request instructing the service to delete an address from the list of verified email addresses.

Request constructor

deleteVerifiedEmailAddress

Arguments

:: Text
dveaEmailAddress

-> DeleteVerifiedEmailAddress

Smart constructor for the minimum required parameters to construct a valid DeleteVerifiedEmailAddress request.

The fields accessible through corresponding lenses are:

EmailAddress :: Text

Request lenses

dveaEmailAddress :: Lens' DeleteVerifiedEmailAddress Text

An email address to be removed from the list of verified addresses.

Response

data DeleteVerifiedEmailAddressResponse

Response constructor

deleteVerifiedEmailAddressResponse :: DeleteVerifiedEmailAddressResponse

Smart constructor for the minimum required parameters to construct a valid DeleteVerifiedEmailAddressResponse response.

This constructor is provided for convenience and testing purposes.