Network.AWS.SES.DescribeActiveReceiptRuleSet

Description

Returns the metadata and receipt rules for the receipt rule set that is currently active.

For information about setting up receipt rule sets, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

describeActiveReceiptRuleSet :: DescribeActiveReceiptRuleSet

Creates a value of DescribeActiveReceiptRuleSet with the minimum fields required to make a request.

data DescribeActiveReceiptRuleSet

Represents a request to return the metadata and receipt rules for the receipt rule set that is currently active. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: describeActiveReceiptRuleSet smart constructor.

Destructuring the Response

describeActiveReceiptRuleSetResponse

Arguments

:: Int
darrsrsResponseStatus

-> DescribeActiveReceiptRuleSetResponse

Creates a value of DescribeActiveReceiptRuleSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darrsrsRules - The receipt rules that belong to the active rule set.

darrsrsMetadata - The metadata for the currently active receipt rule set. The metadata consists of the rule set name and a timestamp of when the rule set was created.

darrsrsResponseStatus - -- | The response status code.

data DescribeActiveReceiptRuleSetResponse

Represents the metadata and receipt rules for the receipt rule set that is currently active.

See: describeActiveReceiptRuleSetResponse smart constructor.

Response Lenses

darrsrsRules :: Lens' DescribeActiveReceiptRuleSetResponse [ReceiptRule]

The receipt rules that belong to the active rule set.

darrsrsMetadata :: Lens' DescribeActiveReceiptRuleSetResponse (Maybe ReceiptRuleSetMetadata)

The metadata for the currently active receipt rule set. The metadata consists of the rule set name and a timestamp of when the rule set was created.

darrsrsResponseStatus :: Lens' DescribeActiveReceiptRuleSetResponse Int

- | The response status code.