Network.AWS.SES.DescribeReceiptRuleSet

Description

Returns the details of the specified receipt rule set.

For information about managing receipt rule sets, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

describeReceiptRuleSet

Arguments

:: Text
drrsRuleSetName

-> DescribeReceiptRuleSet

Creates a value of DescribeReceiptRuleSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drrsRuleSetName - The name of the receipt rule set to describe.

data DescribeReceiptRuleSet

Represents a request to return the details of a receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: describeReceiptRuleSet smart constructor.

Request Lenses

drrsRuleSetName :: Lens' DescribeReceiptRuleSet Text

The name of the receipt rule set to describe.

Destructuring the Response

describeReceiptRuleSetResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeReceiptRuleSetResponse

Creates a value of DescribeReceiptRuleSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsRules - A list of the receipt rules that belong to the specified receipt rule set.

desrsMetadata - The metadata for the receipt rule set, which consists of the rule set name and the timestamp of when the rule set was created.

desrsResponseStatus - -- | The response status code.

data DescribeReceiptRuleSetResponse

Represents the details of the specified receipt rule set.

See: describeReceiptRuleSetResponse smart constructor.

Response Lenses

desrsRules :: Lens' DescribeReceiptRuleSetResponse [ReceiptRule]

A list of the receipt rules that belong to the specified receipt rule set.

desrsMetadata :: Lens' DescribeReceiptRuleSetResponse (Maybe ReceiptRuleSetMetadata)

The metadata for the receipt rule set, which consists of the rule set name and the timestamp of when the rule set was created.

desrsResponseStatus :: Lens' DescribeReceiptRuleSetResponse Int

- | The response status code.