Network.AWS.SES.GetIdentityNotificationAttributes

Description

Given a list of verified identities (email addresses and/or domains), returns a structure describing identity notification attributes.

This action is throttled at one request per second.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_GetIdentityNotificationAttributes.html

Request

data GetIdentityNotificationAttributes

Request constructor

getIdentityNotificationAttributes :: GetIdentityNotificationAttributes

GetIdentityNotificationAttributes constructor.

The fields accessible through corresponding lenses are:

ginaIdentities :: [Text]

Request lenses

ginaIdentities :: Lens' GetIdentityNotificationAttributes [Text]

A list of one or more identities.

Response

data GetIdentityNotificationAttributesResponse

Response constructor

getIdentityNotificationAttributesResponse :: GetIdentityNotificationAttributesResponse

GetIdentityNotificationAttributesResponse constructor.

The fields accessible through corresponding lenses are:

ginarNotificationAttributes :: HashMap Text IdentityNotificationAttributes

Response lenses

ginarNotificationAttributes :: Lens' GetIdentityNotificationAttributesResponse (HashMap Text IdentityNotificationAttributes)

A map of Identity to IdentityNotificationAttributes.