Network.AWS.SES.GetIdentityPolicies

Description

Returns the requested sending authorization policies for the given identity (an email address or a domain). The policies are returned as a map of policy names to policy contents. You can retrieve a maximum of 20 policies at a time.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

getIdentityPolicies

Arguments

:: Text
gipIdentity

-> GetIdentityPolicies

Creates a value of GetIdentityPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gipIdentity - The identity for which the policies will be retrieved. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: userexample.com , example.com , arn:aws:ses:us-east-1:123456789012:identity/example.com@ . To successfully call this API, you must own the identity.

gipPolicyNames - A list of the names of policies to be retrieved. You can retrieve a maximum of 20 policies at a time. If you do not know the names of the policies that are attached to the identity, you can use ListIdentityPolicies .

data GetIdentityPolicies

Represents a request to return the requested sending authorization policies for an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide .

See: getIdentityPolicies smart constructor.

Request Lenses

gipIdentity :: Lens' GetIdentityPolicies Text

The identity for which the policies will be retrieved. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: userexample.com , example.com , arn:aws:ses:us-east-1:123456789012:identity/example.com@ . To successfully call this API, you must own the identity.

gipPolicyNames :: Lens' GetIdentityPolicies [Text]

A list of the names of policies to be retrieved. You can retrieve a maximum of 20 policies at a time. If you do not know the names of the policies that are attached to the identity, you can use ListIdentityPolicies .

Destructuring the Response

getIdentityPoliciesResponse

Arguments

:: Int
giprsResponseStatus

-> GetIdentityPoliciesResponse

Creates a value of GetIdentityPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giprsResponseStatus - -- | The response status code.

giprsPolicies - A map of policy names to policies.

data GetIdentityPoliciesResponse

Represents the requested sending authorization policies.

See: getIdentityPoliciesResponse smart constructor.

Response Lenses

giprsResponseStatus :: Lens' GetIdentityPoliciesResponse Int

- | The response status code.

giprsPolicies :: Lens' GetIdentityPoliciesResponse (HashMap Text Text)

A map of policy names to policies.