Network.AWS.SES.GetIdentityVerificationAttributes

Description

Given a list of identities (email addresses and/or domains), returns the verification status and (for domain identities) the verification token for each identity.

The verification status of an email address is Pending until the email address owner clicks the link within the verification email that Amazon SES sent to that address. If the email address owner clicks the link within 24 hours, the verification status of the email address changes to Success. If the link is not clicked within 24 hours, the verification status changes to "Failed." In that case, if you still want to verify the email address, you must restart the verification process from the beginning.

For domain identities, the domain's verification status is Pending as Amazon SES searches for the required TXT record in the DNS settings of the domain. When Amazon SES detects the record, the domain's verification status changes to Success. If Amazon SES is unable to detect the record within 72 hours, the domain's verification status changes to "Failed." In that case, if you still want to verify the domain, you must restart the verification process from the beginning.

This action is throttled at one request per second and can only get verification attributes for up to 100 identities at a time.

Creating a Request

getIdentityVerificationAttributes :: GetIdentityVerificationAttributes

Creates a value of GetIdentityVerificationAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

givaIdentities - A list of identities.

data GetIdentityVerificationAttributes

Represents a request to return the Amazon SES verification status of a list of identities. For domain identities, this request also returns the verification token. For information about verifying identities with Amazon SES, see the Amazon SES Developer Guide .

See: getIdentityVerificationAttributes smart constructor.

Request Lenses

givaIdentities :: Lens' GetIdentityVerificationAttributes [Text]

A list of identities.

Destructuring the Response

getIdentityVerificationAttributesResponse

Arguments

:: Int
givarsResponseStatus

-> GetIdentityVerificationAttributesResponse

Creates a value of GetIdentityVerificationAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

givarsResponseStatus - -- | The response status code.

givarsVerificationAttributes - A map of Identities to IdentityVerificationAttributes objects.

data GetIdentityVerificationAttributesResponse

The Amazon SES verification status of a list of identities. For domain identities, this response also contains the verification token.

See: getIdentityVerificationAttributesResponse smart constructor.

Response Lenses

givarsResponseStatus :: Lens' GetIdentityVerificationAttributesResponse Int

- | The response status code.

givarsVerificationAttributes :: Lens' GetIdentityVerificationAttributesResponse (HashMap Text IdentityVerificationAttributes)

A map of Identities to IdentityVerificationAttributes objects.