Network.AWS.SES.GetSendQuota

Description

Returns the user's current sending limits.

This action is throttled at one request per second.

http://docs.aws.amazon.com/ses/latest/APIReference/API_GetSendQuota.html

Request

data GetSendQuota

Request constructor

getSendQuota :: GetSendQuota

GetSendQuota constructor.

Response

data GetSendQuotaResponse

Response constructor

getSendQuotaResponse :: GetSendQuotaResponse

GetSendQuotaResponse constructor.

The fields accessible through corresponding lenses are:

gsqrMax24HourSend :: Maybe Double

gsqrMaxSendRate :: Maybe Double

gsqrSentLast24Hours :: Maybe Double

Response lenses

gsqrMax24HourSend :: Lens' GetSendQuotaResponse (Maybe Double)

The maximum number of emails the user is allowed to send in a 24-hour interval.

gsqrMaxSendRate :: Lens' GetSendQuotaResponse (Maybe Double)

The maximum number of emails the user is allowed to send per second.

gsqrSentLast24Hours :: Lens' GetSendQuotaResponse (Maybe Double)

The number of emails sent during the previous 24 hours.