Network.AWS.SES.GetSendStatistics

Description

Returns the user's sending statistics. The result is a list of data points, representing the last two weeks of sending activity.

Each data point in the list contains statistics for a 15-minute interval.

This action is throttled at one request per second.

Creating a Request

getSendStatistics :: GetSendStatistics

Creates a value of GetSendStatistics with the minimum fields required to make a request.

data GetSendStatistics

See: getSendStatistics smart constructor.

Destructuring the Response

getSendStatisticsResponse

Arguments

:: Int
gssrsResponseStatus

-> GetSendStatisticsResponse

Creates a value of GetSendStatisticsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gssrsSendDataPoints - A list of data points, each of which represents 15 minutes of activity.

gssrsResponseStatus - -- | The response status code.

data GetSendStatisticsResponse

Represents a list of data points. This list contains aggregated data from the previous two weeks of your sending activity with Amazon SES.

See: getSendStatisticsResponse smart constructor.

Response Lenses

gssrsSendDataPoints :: Lens' GetSendStatisticsResponse [SendDataPoint]

A list of data points, each of which represents 15 minutes of activity.

gssrsResponseStatus :: Lens' GetSendStatisticsResponse Int

- | The response status code.