Network.AWS.SES.ListIdentities

Description

Returns a list containing all of the identities (email addresses and domains) for a specific AWS Account, regardless of verification status.

This action is throttled at one request per second.

http://docs.aws.amazon.com/ses/latest/APIReference/API_ListIdentities.html

Request

data ListIdentities

Request constructor

listIdentities :: ListIdentities

ListIdentities constructor.

The fields accessible through corresponding lenses are:

liIdentityType :: Maybe IdentityType

liMaxItems :: Maybe Int

liNextToken :: Maybe Text

Request lenses

liIdentityType :: Lens' ListIdentities (Maybe IdentityType)

The type of the identities to list. Possible values are EmailAddress and Domain. If this parameter is omitted, then all identities will be listed.

liMaxItems :: Lens' ListIdentities (Maybe Int)

The maximum number of identities per page. Possible values are 1-100 inclusive.

liNextToken :: Lens' ListIdentities (Maybe Text)

The token to use for pagination.

Response

data ListIdentitiesResponse

Response constructor

listIdentitiesResponse :: ListIdentitiesResponse

ListIdentitiesResponse constructor.

The fields accessible through corresponding lenses are:

lirIdentities :: [Text]

lirNextToken :: Maybe Text

Response lenses

lirIdentities :: Lens' ListIdentitiesResponse [Text]

A list of identities.

lirNextToken :: Lens' ListIdentitiesResponse (Maybe Text)

The token used for pagination.