Network.AWS.SES.ListIdentityPolicies

Description

Returns a list of sending authorization policies that are attached to the given identity (an email address or a domain). This API returns only a list. If you want the actual policy content, you can use GetIdentityPolicies .

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

listIdentityPolicies

Arguments

:: Text
lipIdentity

-> ListIdentityPolicies

Creates a value of ListIdentityPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lipIdentity - The identity that is associated with the policy for which the policies will be listed. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: userexample.com , example.com , arn:aws:ses:us-east-1:123456789012:identity/example.com@ . To successfully call this API, you must own the identity.

data ListIdentityPolicies

Represents a request to return a list of sending authorization policies that are attached to an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide .

See: listIdentityPolicies smart constructor.

Request Lenses

lipIdentity :: Lens' ListIdentityPolicies Text

The identity that is associated with the policy for which the policies will be listed. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: userexample.com , example.com , arn:aws:ses:us-east-1:123456789012:identity/example.com@ . To successfully call this API, you must own the identity.

Destructuring the Response

listIdentityPoliciesResponse

Arguments

:: Int
liprsResponseStatus

-> ListIdentityPoliciesResponse

Creates a value of ListIdentityPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

liprsResponseStatus - -- | The response status code.

liprsPolicyNames - A list of names of policies that apply to the specified identity.

data ListIdentityPoliciesResponse

A list of names of sending authorization policies that apply to an identity.

See: listIdentityPoliciesResponse smart constructor.

Response Lenses

liprsResponseStatus :: Lens' ListIdentityPoliciesResponse Int

- | The response status code.

liprsPolicyNames :: Lens' ListIdentityPoliciesResponse [Text]

A list of names of policies that apply to the specified identity.