Network.AWS.SES.ListReceiptRuleSets

Description

Lists the receipt rule sets that exist under your AWS account. If there are additional receipt rule sets to be retrieved, you will receive a NextToken that you can provide to the next call to ListReceiptRuleSets to retrieve the additional entries.

For information about managing receipt rule sets, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

listReceiptRuleSets :: ListReceiptRuleSets

Creates a value of ListReceiptRuleSets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrrsNextToken - A token returned from a previous call to ListReceiptRuleSets to indicate the position in the receipt rule set list.

data ListReceiptRuleSets

Represents a request to list the receipt rule sets that exist under your AWS account. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: listReceiptRuleSets smart constructor.

Request Lenses

lrrsNextToken :: Lens' ListReceiptRuleSets (Maybe Text)

A token returned from a previous call to ListReceiptRuleSets to indicate the position in the receipt rule set list.

Destructuring the Response

listReceiptRuleSetsResponse

Arguments

:: Int
lrrsrsResponseStatus

-> ListReceiptRuleSetsResponse

Creates a value of ListReceiptRuleSetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrrsrsRuleSets - The metadata for the currently active receipt rule set. The metadata consists of the rule set name and the timestamp of when the rule set was created.

lrrsrsNextToken - A token indicating that there are additional receipt rule sets available to be listed. Pass this token to successive calls of ListReceiptRuleSets to retrieve up to 100 receipt rule sets at a time.

lrrsrsResponseStatus - -- | The response status code.

data ListReceiptRuleSetsResponse

A list of receipt rule sets that exist under your AWS account.

See: listReceiptRuleSetsResponse smart constructor.

Response Lenses

lrrsrsRuleSets :: Lens' ListReceiptRuleSetsResponse [ReceiptRuleSetMetadata]

The metadata for the currently active receipt rule set. The metadata consists of the rule set name and the timestamp of when the rule set was created.

lrrsrsNextToken :: Lens' ListReceiptRuleSetsResponse (Maybe Text)

A token indicating that there are additional receipt rule sets available to be listed. Pass this token to successive calls of ListReceiptRuleSets to retrieve up to 100 receipt rule sets at a time.

lrrsrsResponseStatus :: Lens' ListReceiptRuleSetsResponse Int

- | The response status code.