Network.AWS.SES.ListVerifiedEmailAddresses

Description

Returns a list containing all of the email addresses that have been verified.

The ListVerifiedEmailAddresses action is deprecated as of the May 15, 2012 release of Domain Verification. The ListIdentities action is now preferred. This action is throttled at one request per second.

http://docs.aws.amazon.com/ses/latest/APIReference/API_ListVerifiedEmailAddresses.html

Request

data ListVerifiedEmailAddresses

Request constructor

listVerifiedEmailAddresses :: ListVerifiedEmailAddresses

ListVerifiedEmailAddresses constructor.

Response

data ListVerifiedEmailAddressesResponse

Response constructor

listVerifiedEmailAddressesResponse :: ListVerifiedEmailAddressesResponse

ListVerifiedEmailAddressesResponse constructor.

The fields accessible through corresponding lenses are:

lvearVerifiedEmailAddresses :: [Text]

Response lenses

lvearVerifiedEmailAddresses :: Lens' ListVerifiedEmailAddressesResponse [Text]

A list of email addresses that have been verified.