Network.AWS.SES.ReorderReceiptRuleSet

Description

Reorders the receipt rules within a receipt rule set.

For information about managing receipt rule sets, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

reorderReceiptRuleSet

Arguments

:: Text
rrrsRuleSetName

-> ReorderReceiptRuleSet

Creates a value of ReorderReceiptRuleSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrrsRuleSetName - The name of the receipt rule set to reorder.

rrrsRuleNames - A list of the specified receipt rule set's receipt rules in the order that you want to put them.

data ReorderReceiptRuleSet

Represents a request to reorder the receipt rules within a receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: reorderReceiptRuleSet smart constructor.

Request Lenses

rrrsRuleSetName :: Lens' ReorderReceiptRuleSet Text

The name of the receipt rule set to reorder.

rrrsRuleNames :: Lens' ReorderReceiptRuleSet [Text]

A list of the specified receipt rule set's receipt rules in the order that you want to put them.

Destructuring the Response

reorderReceiptRuleSetResponse

Arguments

:: Int
rrrsrsResponseStatus

-> ReorderReceiptRuleSetResponse

Creates a value of ReorderReceiptRuleSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrrsrsResponseStatus - -- | The response status code.

data ReorderReceiptRuleSetResponse

An empty element returned on a successful request.

See: reorderReceiptRuleSetResponse smart constructor.

Response Lenses

rrrsrsResponseStatus :: Lens' ReorderReceiptRuleSetResponse Int

- | The response status code.