Network.AWS.SES.SendBounce

Description

Generates and sends a bounce message to the sender of an email you received through Amazon SES. You can only use this API on an email up to 24 hours after you receive it.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

sendBounce

Arguments

:: Text
sbOriginalMessageId

-> Text
sbBounceSender

-> SendBounce

Creates a value of SendBounce with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sbMessageDsn - Message-related DSN fields. If not specified, Amazon SES will choose the values.

sbExplanation - Human-readable text for the bounce message to explain the failure. If not specified, the text will be auto-generated based on the bounced recipient information.

sbBounceSenderARN - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the address in the From header of the bounce. For more information about sending authorization, see the Amazon SES Developer Guide .

sbOriginalMessageId - The message ID of the message to be bounced.

sbBounceSender - The address to use in the From header of the bounce message. This must be an identity that you have verified with Amazon SES.

sbBouncedRecipientInfoList - A list of recipients of the bounced message, including the information required to create the Delivery Status Notifications (DSNs) for the recipients. You must specify at least one BouncedRecipientInfo in the list.

data SendBounce

Represents a request to send a bounce message to the sender of an email you received through Amazon SES.

See: sendBounce smart constructor.

Request Lenses

sbMessageDsn :: Lens' SendBounce (Maybe MessageDsn)

Message-related DSN fields. If not specified, Amazon SES will choose the values.

sbExplanation :: Lens' SendBounce (Maybe Text)

Human-readable text for the bounce message to explain the failure. If not specified, the text will be auto-generated based on the bounced recipient information.

sbBounceSenderARN :: Lens' SendBounce (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the address in the From header of the bounce. For more information about sending authorization, see the Amazon SES Developer Guide .

sbOriginalMessageId :: Lens' SendBounce Text

The message ID of the message to be bounced.

sbBounceSender :: Lens' SendBounce Text

The address to use in the From header of the bounce message. This must be an identity that you have verified with Amazon SES.

sbBouncedRecipientInfoList :: Lens' SendBounce [BouncedRecipientInfo]

A list of recipients of the bounced message, including the information required to create the Delivery Status Notifications (DSNs) for the recipients. You must specify at least one BouncedRecipientInfo in the list.

Destructuring the Response

sendBounceResponse

Arguments

:: Int
sbrsResponseStatus

-> SendBounceResponse

Creates a value of SendBounceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sbrsMessageId - The message ID of the bounce message.

sbrsResponseStatus - -- | The response status code.

data SendBounceResponse

Represents a unique message ID.

See: sendBounceResponse smart constructor.

Response Lenses

sbrsMessageId :: Lens' SendBounceResponse (Maybe Text)

The message ID of the bounce message.

sbrsResponseStatus :: Lens' SendBounceResponse Int

- | The response status code.