Network.AWS.SES.SendEmail

Description

Composes an email message based on input data, and then immediately queues the message for sending.

You can only send email from verified email addresses and domains. If you have not requested production access to Amazon SES, you must also verify every recipient email address except for the recipients provided by the Amazon SES mailbox simulator. For more information, go to the Amazon SESDeveloper Guide. The total size of the message cannot exceed 10 MB.

Amazon SES has a limit on the total number of recipients per message: The combined number of To:, CC: and BCC: email addresses cannot exceed 50. If you need to send an email message to a larger audience, you can divide your recipient list into groups of 50 or fewer, and then call Amazon SES repeatedly to send the message to each group.

For every message that you send, the total number of recipients (To:, CC: and BCC:) is counted against your sending quota - the maximum number of emails you can send in a 24-hour period. For information about your sending quota, go to the Amazon SES Developer Guide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_SendEmail.html

Request

data SendEmail

Request constructor

sendEmail

Arguments

:: Text
seSource

-> Destination
seDestination

-> Message
seMessage

-> SendEmail

SendEmail constructor.

The fields accessible through corresponding lenses are:

seDestination :: Destination

seMessage :: Message

seReplyToAddresses :: [Text]

seReturnPath :: Maybe Text

seSource :: Text

Request lenses

seDestination :: Lens' SendEmail Destination

The destination for this email, composed of To:, CC:, and BCC: fields.

seMessage :: Lens' SendEmail Message

The message to be sent.

seReplyToAddresses :: Lens' SendEmail [Text]

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

seReturnPath :: Lens' SendEmail (Maybe Text)

The email address to which bounces and complaints are to be forwarded when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter.

seSource :: Lens' SendEmail Text

The identity's email address.

By default, the string must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: '=?charset?encoding?encoded-text?='. For more information, see RFC 2047.

Response

data SendEmailResponse

Response constructor

sendEmailResponse

Arguments

:: Text
serMessageId

-> SendEmailResponse

SendEmailResponse constructor.

The fields accessible through corresponding lenses are:

serMessageId :: Text

Response lenses

serMessageId :: Lens' SendEmailResponse Text

The unique message identifier returned from the SendEmail action.