Network.AWS.SES.SetIdentityDkimEnabled

Description

Enables or disables Easy DKIM signing of email sent from an identity:

If Easy DKIM signing is enabled for a domain name identity (e.g., 'example.com'), then Amazon SES will DKIM-sign all email sent by addresses under that domain name (e.g., 'user@example.com'). If Easy DKIM signing is enabled for an email address, then Amazon SES will DKIM-sign all email sent by that email address. For email addresses (e.g., 'user@example.com'), you can only enable Easy DKIM signing if the corresponding domain (e.g., 'example.com') has been set up for Easy DKIM using the AWS Console or the VerifyDomainDkim action.

This action is throttled at one request per second.

For more information about Easy DKIM signing, go to the Amazon SES DeveloperGuide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_SetIdentityDkimEnabled.html

Request

data SetIdentityDkimEnabled

Request constructor

setIdentityDkimEnabled

Arguments

:: Text
sideIdentity

-> Bool
sideDkimEnabled

-> SetIdentityDkimEnabled

SetIdentityDkimEnabled constructor.

The fields accessible through corresponding lenses are:

sideDkimEnabled :: Bool

sideIdentity :: Text

Request lenses

sideDkimEnabled :: Lens' SetIdentityDkimEnabled Bool

Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

sideIdentity :: Lens' SetIdentityDkimEnabled Text

The identity for which DKIM signing should be enabled or disabled.

Response

data SetIdentityDkimEnabledResponse

Response constructor

setIdentityDkimEnabledResponse :: SetIdentityDkimEnabledResponse

SetIdentityDkimEnabledResponse constructor.