Network.AWS.SES.SetIdentityHeadersInNotificationsEnabled

Description

Given an identity (an email address or a domain), sets whether Amazon SES includes the original email headers in the Amazon Simple Notification Service (Amazon SNS) notifications of a specified type.

This action is throttled at one request per second.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide .

Creating a Request

setIdentityHeadersInNotificationsEnabled

Arguments

:: Text
sihineIdentity

-> NotificationType
sihineNotificationType

-> Bool
sihineEnabled

-> SetIdentityHeadersInNotificationsEnabled

Creates a value of SetIdentityHeadersInNotificationsEnabled with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sihineIdentity - The identity for which to enable or disable headers in notifications. Examples: userexample.com , example.com@ .

sihineNotificationType - The notification type for which to enable or disable headers in notifications.

sihineEnabled - Sets whether Amazon SES includes the original email headers in Amazon SNS notifications of the specified notification type. A value of true specifies that Amazon SES will include headers in notifications, and a value of false specifies that Amazon SES will not include headers in notifications. This value can only be set when NotificationType is already set to use a particular Amazon SNS topic.

data SetIdentityHeadersInNotificationsEnabled

Represents a request to set whether Amazon SES includes the original email headers in the Amazon SNS notifications of a specified type. For information about notifications, see the Amazon SES Developer Guide .

See: setIdentityHeadersInNotificationsEnabled smart constructor.

Request Lenses

sihineIdentity :: Lens' SetIdentityHeadersInNotificationsEnabled Text

The identity for which to enable or disable headers in notifications. Examples: userexample.com , example.com@ .

sihineNotificationType :: Lens' SetIdentityHeadersInNotificationsEnabled NotificationType

The notification type for which to enable or disable headers in notifications.

sihineEnabled :: Lens' SetIdentityHeadersInNotificationsEnabled Bool

Sets whether Amazon SES includes the original email headers in Amazon SNS notifications of the specified notification type. A value of true specifies that Amazon SES will include headers in notifications, and a value of false specifies that Amazon SES will not include headers in notifications. This value can only be set when NotificationType is already set to use a particular Amazon SNS topic.

Destructuring the Response

setIdentityHeadersInNotificationsEnabledResponse

Arguments

:: Int
sihinersResponseStatus

-> SetIdentityHeadersInNotificationsEnabledResponse

Creates a value of SetIdentityHeadersInNotificationsEnabledResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sihinersResponseStatus - -- | The response status code.

data SetIdentityHeadersInNotificationsEnabledResponse

An empty element returned on a successful request.

See: setIdentityHeadersInNotificationsEnabledResponse smart constructor.

Response Lenses

sihinersResponseStatus :: Lens' SetIdentityHeadersInNotificationsEnabledResponse Int

- | The response status code.