Network.AWS.SES.SetIdentityMailFromDomain

Description

Enables or disables the custom MAIL FROM domain setup for a verified identity (an email address or a domain).

Important: To send emails using the specified MAIL FROM domain, you must add an MX record to your MAIL FROM domain's DNS settings. If you want your emails to pass Sender Policy Framework (SPF) checks, you must also add or update an SPF record. For more information, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

setIdentityMailFromDomain

Arguments

:: Text
simfdIdentity

-> SetIdentityMailFromDomain

Creates a value of SetIdentityMailFromDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

simfdMailFromDomain - The custom MAIL FROM domain that you want the verified identity to use. The MAIL FROM domain must 1) be a subdomain of the verified identity, 2) not be used in a From address if the MAIL FROM domain is the destination of email feedback forwarding (for more information, see the Amazon SES Developer Guide ), and 3) not be used to receive emails. A value of null disables the custom MAIL FROM setting for the identity.

simfdBehaviorOnMXFailure - The action that you want Amazon SES to take if it cannot successfully read the required MX record when you send an email. If you choose UseDefaultValue , Amazon SES will use amazonses.com (or a subdomain of that) as the MAIL FROM domain. If you choose RejectMessage , Amazon SES will return a MailFromDomainNotVerified error and not send the email. The action specified in BehaviorOnMXFailure is taken when the custom MAIL FROM domain setup is in the Pending , Failed , and TemporaryFailure states.

simfdIdentity - The verified identity for which you want to enable or disable the specified custom MAIL FROM domain.

data SetIdentityMailFromDomain

Represents a request to enable or disable the Amazon SES custom MAIL FROM domain setup for a verified identity. For information about using a custom MAIL FROM domain, see the Amazon SES Developer Guide .

See: setIdentityMailFromDomain smart constructor.

Request Lenses

simfdMailFromDomain :: Lens' SetIdentityMailFromDomain (Maybe Text)

The custom MAIL FROM domain that you want the verified identity to use. The MAIL FROM domain must 1) be a subdomain of the verified identity, 2) not be used in a From address if the MAIL FROM domain is the destination of email feedback forwarding (for more information, see the Amazon SES Developer Guide ), and 3) not be used to receive emails. A value of null disables the custom MAIL FROM setting for the identity.

simfdBehaviorOnMXFailure :: Lens' SetIdentityMailFromDomain (Maybe BehaviorOnMXFailure)

The action that you want Amazon SES to take if it cannot successfully read the required MX record when you send an email. If you choose UseDefaultValue , Amazon SES will use amazonses.com (or a subdomain of that) as the MAIL FROM domain. If you choose RejectMessage , Amazon SES will return a MailFromDomainNotVerified error and not send the email. The action specified in BehaviorOnMXFailure is taken when the custom MAIL FROM domain setup is in the Pending , Failed , and TemporaryFailure states.

simfdIdentity :: Lens' SetIdentityMailFromDomain Text

The verified identity for which you want to enable or disable the specified custom MAIL FROM domain.

Destructuring the Response

setIdentityMailFromDomainResponse

Arguments

:: Int
simfdrsResponseStatus

-> SetIdentityMailFromDomainResponse

Creates a value of SetIdentityMailFromDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

simfdrsResponseStatus - -- | The response status code.

data SetIdentityMailFromDomainResponse

An empty element returned on a successful request.

See: setIdentityMailFromDomainResponse smart constructor.

Response Lenses

simfdrsResponseStatus :: Lens' SetIdentityMailFromDomainResponse Int

- | The response status code.