Network.AWS.SES.SetIdentityNotificationTopic

Description

Given an identity (email address or domain), sets the Amazon Simple Notification Service (Amazon SNS) topic to which Amazon SES will publish bounce, complaint, and/or delivery notifications for emails sent with that identity as the Source.

This action is throttled at one request per second.

For more information about feedback notification, see the Amazon SESDeveloper Guide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_SetIdentityNotificationTopic.html

Request

data SetIdentityNotificationTopic

Request constructor

setIdentityNotificationTopic

Arguments

:: Text
sintIdentity

-> NotificationType
sintNotificationType

-> SetIdentityNotificationTopic

SetIdentityNotificationTopic constructor.

The fields accessible through corresponding lenses are:

sintIdentity :: Text

sintNotificationType :: NotificationType

sintSnsTopic :: Maybe Text

Request lenses

sintIdentity :: Lens' SetIdentityNotificationTopic Text

The identity for which the Amazon SNS topic will be set. Examples: 'user@example.com', 'example.com'.

sintNotificationType :: Lens' SetIdentityNotificationTopic NotificationType

The type of notifications that will be published to the specified Amazon SNS topic.

sintSnsTopic :: Lens' SetIdentityNotificationTopic (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic. If the parameter is omitted from the request or a null value is passed, SnsTopic is cleared and publishing is disabled.

Response

data SetIdentityNotificationTopicResponse

Response constructor

setIdentityNotificationTopicResponse :: SetIdentityNotificationTopicResponse

SetIdentityNotificationTopicResponse constructor.