Network.AWS.SES.SetReceiptRulePosition

Description

Sets the position of the specified receipt rule in the receipt rule set.

For information about managing receipt rules, see the Amazon SES Developer Guide .

This action is throttled at one request per second.

Creating a Request

setReceiptRulePosition

Arguments

:: Text
srrpRuleSetName

-> Text
srrpRuleName

-> SetReceiptRulePosition

Creates a value of SetReceiptRulePosition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srrpAfter - The name of the receipt rule after which to place the specified receipt rule.

srrpRuleSetName - The name of the receipt rule set that contains the receipt rule to reposition.

srrpRuleName - The name of the receipt rule to reposition.

data SetReceiptRulePosition

Represents a request to set the position of a receipt rule in a receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide .

See: setReceiptRulePosition smart constructor.

Request Lenses

srrpAfter :: Lens' SetReceiptRulePosition (Maybe Text)

The name of the receipt rule after which to place the specified receipt rule.

srrpRuleSetName :: Lens' SetReceiptRulePosition Text

The name of the receipt rule set that contains the receipt rule to reposition.

srrpRuleName :: Lens' SetReceiptRulePosition Text

The name of the receipt rule to reposition.

Destructuring the Response

setReceiptRulePositionResponse

Arguments

:: Int
srrprsResponseStatus

-> SetReceiptRulePositionResponse

Creates a value of SetReceiptRulePositionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srrprsResponseStatus - -- | The response status code.

data SetReceiptRulePositionResponse

An empty element returned on a successful request.

See: setReceiptRulePositionResponse smart constructor.

Response Lenses

srrprsResponseStatus :: Lens' SetReceiptRulePositionResponse Int

- | The response status code.