Network.AWS.SES.V2010_12_01.DeleteIdentity

Description

Deletes the specified identity (email address or domain) from the list of verified identities. This action is throttled at one request per second. POST HTTP1.1 Date: Sat, 12 May 2012 05:25:58 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=w943pl3zIvtszwzZxypi+LsgjzquvhYhnG42S6b2WLo=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 135 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=DeleteIdentity &Identity=domain.com &Timestamp=2012-05-12T05%3A25%3A58.000Z &Version=2010-12-01 d96bd874-9bf2-11e1-8ee7-c98a0037a2b6.

Request

data DeleteIdentity

Represents a request instructing the service to delete an identity from the list of identities for the AWS Account.

Request constructor

mkDeleteIdentity

Arguments

:: Text
diIdentity

-> DeleteIdentity

Smart constructor for the minimum required parameters to construct a valid DeleteIdentity request.

Request lenses

diIdentity :: Lens' DeleteIdentity Text

The identity to be removed from the list of identities for the AWS Account.

Response

data DeleteIdentityResponse

An empty element. Receiving this element indicates that the request completed successfully.