Network.AWS.SES.V2010_12_01.DeleteVerifiedEmailAddress

Description

Deletes the specified email address from the list of verified addresses. The DeleteVerifiedEmailAddress action is deprecated as of the May 15, 2012 release of Domain Verification. The DeleteIdentity action is now preferred. This action is throttled at one request per second. POST HTTP1.1 Date: Thu, 18 Aug 2011 22:20:50 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=Rxzyd6cQe/YjkV4yoQAZ243OzzNjFgrsclizTKwRIRc=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 142 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=DeleteVerifiedEmailAddress &EmailAddress=user%40example.com &Timestamp=2011-08-18T22%3A20%3A50.000Z 5634af08-c865-11e0-8986-3f99a698f914.

Documentation

data DeleteVerifiedEmailAddress

Constructors

DeleteVerifiedEmailAddress

Fields

_dvearEmailAddress :: Text

An email address to be removed from the list of verified addresses.

dvearEmailAddress :: Iso' DeleteVerifiedEmailAddress Text

data DeleteVerifiedEmailAddressResponse

Constructors

DeleteVerifiedEmailAddressResponse