Network.AWS.SES.V2010_12_01.GetIdentityNotificationAttributes

Description

Given a list of verified identities (email addresses and/or domains), returns a structure describing identity notification attributes. This action is throttled at one request per second. For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide. POST HTTP1.1 Date: Fri, 15 Jun 2012 20:51:42 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=ee9aH6tUW5wBPoh01Tz3w4H+z4avrMmvmRYbfORC7OI=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 173 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=GetIdentityNotificationAttributes &Identities.member.1=user%40example.com &Timestamp=2012-06-15T20%3A51%3A42.000Z &Version=2010-12-01 user@example.com true arn:aws:sns:us-east-1:123456789012:example arn:aws:sns:us-east-1:123456789012:example arn:aws:sns:us-east-1:123456789012:example e038e509-b72a-11e1-901f-1fbd90e8104f.

Request

data GetIdentityNotificationAttributes

Returns the notification attributes for a given list of identities (email address or domain names).

Request constructor

mkGetIdentityNotificationAttributes

Arguments

:: [Text]
ginaIdentities

-> GetIdentityNotificationAttributes

Smart constructor for the minimum required parameters to construct a valid GetIdentityNotificationAttributes request.

Request lenses

ginaIdentities :: Lens' GetIdentityNotificationAttributes [Text]

A list of one or more identities.

Response

data GetIdentityNotificationAttributesResponse

Describes whether an identity has Amazon Simple Notification Service (Amazon SNS) topics set for bounce, complaint, and/or delivery notifications, and specifies whether feedback forwarding is enabled for bounce and complaint notifications.

Response lenses

ginarNotificationAttributes :: Lens' GetIdentityNotificationAttributesResponse (Map Text IdentityNotificationAttributes)

A map of Identity to IdentityNotificationAttributes.