Network.AWS.SES.V2010_12_01.GetIdentityVerificationAttributes

Description

Given a list of identities (email addresses and/or domains), returns the verification status and (for domain identities) the verification token for each identity. This action is throttled at one request per second. POST / HTTP/1.1 Date: Sat, 12 May 2012 05:27:54 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=3+KQ4VHx991T7Kb41HmFcZJxuHz4/6mf2H5FxY+tuLc=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 203 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=GetIdentityVerificationAttributes &Identities.member.1=user%40domain.com &Identities.member.2=domain.com &Timestamp=2012-05-12T05%3A27%3A54.000Z &Version=2010-12-01 domain.com Pending QTKknzFg2J4ygwa+XvHAxUl1hyHoY0gVfZdfjIedHZ0= user@domain.com Pending 1d0c29f1-9bf3-11e1-8ee7-c98a0037a2b6.

Request

data GetIdentityVerificationAttributes

Represents a request instructing the service to provide the verification attributes for a list of identities.

Request constructor

mkGetIdentityVerificationAttributes

Arguments

:: [Text]
givaIdentities

-> GetIdentityVerificationAttributes

Smart constructor for the minimum required parameters to construct a valid GetIdentityVerificationAttributes request.

Request lenses

givaIdentities :: Lens' GetIdentityVerificationAttributes [Text]

A list of identities.

Response

data GetIdentityVerificationAttributesResponse

Represents the verification attributes for a list of identities.

Response lenses

givarVerificationAttributes :: Lens' GetIdentityVerificationAttributesResponse (Map Text IdentityVerificationAttributes)

A map of Identities to IdentityVerificationAttributes objects.