Network.AWS.SES.V2010_12_01.GetSendStatistics

Description

Returns the user's sending statistics. The result is a list of data points, representing the last two weeks of sending activity. Each data point in the list contains statistics for a 15-minute interval. This action is throttled at one request per second. POST HTTP1.1 Date: Thu, 18 Aug 2011 22:23:01 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=kwuk4eraA9HSfHySflgDKR6xK0JXjATIE7Uu5/FB4x4=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 99 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=GetSendStatistics &Timestamp=2011-08-18T22%3A23%3A01.000Z 8 2011-08-03T19:23:00Z 0 0 0 7 2011-08-03T06:53:00Z 0 0 0 . . . . c2b66ee5-c866-11e0-b17f-cddb0ab334db.

Documentation

data GetSendStatistics

Constructors

GetSendStatistics

data GetSendStatisticsResponse

Constructors

GetSendStatisticsResponse

Fields

_gssrSendDataPoints :: [SendDataPoint]

A list of data points, each of which represents 15 minutes of activity.

gssrSendDataPoints :: Iso' GetSendStatisticsResponse [SendDataPoint]