Network.AWS.SES.V2010_12_01.SetIdentityDkimEnabled

Description

Enables or disables Easy DKIM signing of email sent from an identity: If Easy DKIM signing is enabled for a domain name identity (e.g., example.com), then Amazon SES will DKIM-sign all email sent by addresses under that domain name (e.g., userexample.com). If Easy DKIM signing is enabled for an email address, then Amazon SES will DKIM-sign all email sent by that email address. For email addresses (e.g., userexample.com), you can only enable Easy DKIM signing if the corresponding domain (e.g., example.com) has been set up for Easy DKIM using the AWS Console or the VerifyDomainDkim action. This action is throttled at one request per second. For more information about Easy DKIM signing, go to the Amazon SES Developer Guide. POST HTTP1.1 Date: Fri, 29 Jun 2012 22:42:08 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=u/hDNhYm87AV7LAPzouTBz6HJxUEuE5k96sLzYHjR24=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 168 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=SetIdentityDkimEnabled &DkimEnabled=true&Identity=user%40example.com &Timestamp=2012-06-29T22%3A42%3A08.000Z &Version=2010-12-01 7aa61362-c469-11e1-aee5-6bbb4608fbcc.

Documentation

data SetIdentityDkimEnabled

Constructors

SetIdentityDkimEnabled

Fields

_siderDkimEnabled :: Bool

Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

_siderIdentity :: Text

The identity for which DKIM signing should be enabled or disabled.

siderIdentity :: Lens' SetIdentityDkimEnabled Text

siderDkimEnabled :: Lens' SetIdentityDkimEnabled Bool

data SetIdentityDkimEnabledResponse

Constructors

SetIdentityDkimEnabledResponse