Network.AWS.SES.V2010_12_01.SetIdentityNotificationTopic

Description

Given an identity (email address or domain), sets the Amazon Simple Notification Service (Amazon SNS) topic to which Amazon SES will publish bounce, complaint, and/or delivery notifications for emails sent with that identity as the Source. Unless feedback forwarding is enabled, you must specify Amazon SNS topics for bounce and complaint notifications. For more information, see SetIdentityFeedbackForwardingEnabled. This action is throttled at one request per second. For more information about feedback notification, see the Amazon SES Developer Guide. POST HTTP1.1 Date: Sat, 12 May 2012 05:27:54 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=3+KQ4VHx991T7Kb41HmFcZJxuHz4/6mf2H5FxY+tuLc=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 203 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=SetIdentityNotificationTopic &Identity=user@example.com &SnsTopic=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3Aexample &NotificationType=Bounce &Timestamp=2012-05-12T05%3A27%3A54.000Z&Version=2010-12-01 299f4af4-b72a-11e1-901f-1fbd90e8104f.

Request

data SetIdentityNotificationTopic

Represents a request to set or clear an identity's notification topic.

Request constructor

mkSetIdentityNotificationTopic

Arguments

:: Text
sintIdentity

-> NotificationType
sintNotificationType

-> SetIdentityNotificationTopic

Smart constructor for the minimum required parameters to construct a valid SetIdentityNotificationTopic request.

The fields accessible through corresponding lenses are:

Identity :: Text

NotificationType :: NotificationType

SnsTopic :: Maybe Text

Request lenses

sintIdentity :: Lens' SetIdentityNotificationTopic Text

The identity for which the Amazon SNS topic will be set. Examples: user@example.com, example.com.

sintNotificationType :: Lens' SetIdentityNotificationTopic NotificationType

The type of notifications that will be published to the specified Amazon SNS topic.

sintSnsTopic :: Lens' SetIdentityNotificationTopic (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic. If the parameter is omitted from the request or a null value is passed, SnsTopic is cleared and publishing is disabled.

Response

data SetIdentityNotificationTopicResponse

An empty element. Receiving this element indicates that the request completed successfully.

Response constructor

mkSetIdentityNotificationTopicResponse :: SetIdentityNotificationTopicResponse

Smart constructor for the minimum required parameters to construct a valid SetIdentityNotificationTopicResponse response.

This constructor is provided for convenience and testing purposes.