Network.AWS.SES.V2010_12_01.Types

Description

Amazon Simple Email Service (Amazon SES) is a cost-effective outbound-only email-sending service built on the reliable and scalable infrastructure that Amazon.com has developed to serve its own customer base. With Amazon SES, you can send transactional email, marketing messages, or any other type of high-quality content and you only pay for what you use. Along with high deliverability, Amazon SES provides easy, real-time access to your sending statistics and built-in notifications for bounces, complaints, and deliveries to help you fine-tune your email-sending strategy.

Service

data SES

Supported version (2010-12-01) of the Amazon Simple Email Service service.

XML

xmlOptions :: Tagged a XMLOptions

IdentityType

data IdentityType

Constructors

IdentityTypeDomain
Domain

IdentityTypeEmailAddress
EmailAddress

NotificationType

data NotificationType

Constructors

NotificationTypeBounce
Bounce

NotificationTypeComplaint
Complaint

NotificationTypeDelivery
Delivery

VerificationStatus

data VerificationStatus

Constructors

VerificationStatusFailed
Failed

VerificationStatusNotStarted
NotStarted

VerificationStatusPending
Pending

VerificationStatusSuccess
Success

VerificationStatusTemporaryFailure
TemporaryFailure

RawMessage

data RawMessage

The raw text of the message. The client is responsible for ensuring the following: Message must contain a header and a body, separated by a blank line. All required header fields must be present. Each part of a multipart MIME message must be formatted properly. MIME content types must be among those supported by Amazon SES. For more information, go to the Amazon SES Developer Guide. Content must be base64-encoded, if MIME requires it.

mkRawMessage

Arguments

:: ByteString
rmData

-> RawMessage

Smart constructor for the minimum required fields to construct a valid RawMessage data type to populate a request.

rmData :: Lens' RawMessage ByteString

The raw data of the message. The client must ensure that the message format complies with Internet email standards regarding email header fields, MIME types, MIME encoding, and base64 encoding (if necessary). The To:, CC:, and BCC: headers in the raw message can contain a group list. For more information, go to the Amazon SES Developer Guide.

Body

data Body

The message body.

mkBody :: Body

Smart constructor for the minimum required fields to construct a valid Body data type to populate a request.

bText :: Lens' Body (Maybe Content)

The content of the message, in text format. Use this for text-based email clients, or clients on high-latency networks (such as mobile devices).

bHtml :: Lens' Body (Maybe Content)

The content of the message, in HTML format. Use this for email clients that can process HTML. You can include clickable links, formatted text, and much more in an HTML message.

Content

data Content

The subject of the message: A short summary of the content, which will appear in the recipient's inbox.

mkContent

Arguments

:: Text
cData

-> Content

Smart constructor for the minimum required fields to construct a valid Content data type to populate a request.

cData :: Lens' Content Text

The textual data of the content.

cCharset :: Lens' Content (Maybe Text)

The character set of the content.

Destination

data Destination

The destination for this email, composed of To:, CC:, and BCC: fields.

mkDestination :: Destination

Smart constructor for the minimum required fields to construct a valid Destination data type to populate a request.

dToAddresses :: Lens' Destination [Text]

The To: field(s) of the message.

dCcAddresses :: Lens' Destination [Text]

The CC: field(s) of the message.

dBccAddresses :: Lens' Destination [Text]

The BCC: field(s) of the message.

IdentityDkimAttributes

data IdentityDkimAttributes

Represents the DKIM attributes of a verified email address or a domain.

mkIdentityDkimAttributes

Arguments

:: Bool
idaDkimEnabled

-> VerificationStatus
idaDkimVerificationStatus

-> IdentityDkimAttributes

Smart constructor for the minimum required fields to construct a valid IdentityDkimAttributes data type.

IdentityDkimAttributes is exclusively used in responses and this constructor is provided for convenience and testing purposes.

idaDkimEnabled :: Lens' IdentityDkimAttributes Bool

True if DKIM signing is enabled for email sent from the identity; false otherwise.

idaDkimVerificationStatus :: Lens' IdentityDkimAttributes VerificationStatus

Describes whether Amazon SES has successfully verified the DKIM DNS records (tokens) published in the domain name's DNS. (This only applies to domain identities, not email address identities.).

idaDkimTokens :: Lens' IdentityDkimAttributes [Text]

A set of character strings that represent the domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain. (This only applies to domain identities, not email address identities.) For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

IdentityNotificationAttributes

data IdentityNotificationAttributes

Represents the notification attributes of an identity, including whether an identity has Amazon Simple Notification Service (Amazon SNS) topics set for bounce, complaint, and/or delivery notifications, and whether feedback forwarding is enabled for bounce and complaint notifications.

mkIdentityNotificationAttributes

Arguments

:: Text
inaBounceTopic

-> Text
inaComplaintTopic

-> Text
inaDeliveryTopic

-> Bool
inaForwardingEnabled

-> IdentityNotificationAttributes

Smart constructor for the minimum required fields to construct a valid IdentityNotificationAttributes data type.

IdentityNotificationAttributes is exclusively used in responses and this constructor is provided for convenience and testing purposes.

inaBounceTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish bounce notifications.

inaComplaintTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish complaint notifications.

inaDeliveryTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish delivery notifications.

inaForwardingEnabled :: Lens' IdentityNotificationAttributes Bool

Describes whether Amazon SES will forward bounce and complaint notifications as email. true indicates that Amazon SES will forward bounce and complaint notifications as email, while false indicates that bounce and complaint notifications will be published only to the specified bounce and complaint Amazon SNS topics.

IdentityVerificationAttributes

data IdentityVerificationAttributes

Represents the verification attributes of a single identity.

mkIdentityVerificationAttributes

Arguments

:: VerificationStatus
ivaVerificationStatus

-> IdentityVerificationAttributes

Smart constructor for the minimum required fields to construct a valid IdentityVerificationAttributes data type.

IdentityVerificationAttributes is exclusively used in responses and this constructor is provided for convenience and testing purposes.

ivaVerificationStatus :: Lens' IdentityVerificationAttributes VerificationStatus

The verification status of the identity: Pending, Success, Failed, or TemporaryFailure.

ivaVerificationToken :: Lens' IdentityVerificationAttributes (Maybe Text)

The verification token for a domain identity. Null for email address identities.

Message

data Message

The message to be sent.

mkMessage

Arguments

:: Content
mSubject

-> Body
mBody

-> Message

Smart constructor for the minimum required fields to construct a valid Message data type to populate a request.

mSubject :: Lens' Message Content

The subject of the message: A short summary of the content, which will appear in the recipient's inbox.

mBody :: Lens' Message Body

The message body.

SendDataPoint

data SendDataPoint

Represents sending statistics data. Each SendDataPoint contains statistics for a 15-minute period of sending activity.

mkSendDataPoint :: SendDataPoint

Smart constructor for the minimum required fields to construct a valid SendDataPoint data type.

SendDataPoint is exclusively used in responses and this constructor is provided for convenience and testing purposes.

sdpTimestamp :: Lens' SendDataPoint (Maybe ISO8601)

Time of the data point.

sdpDeliveryAttempts :: Lens' SendDataPoint (Maybe Integer)

Number of emails that have been enqueued for sending.

sdpBounces :: Lens' SendDataPoint (Maybe Integer)

Number of emails that have bounced.

sdpComplaints :: Lens' SendDataPoint (Maybe Integer)

Number of unwanted emails that were rejected by recipients.

sdpRejects :: Lens' SendDataPoint (Maybe Integer)

Number of emails rejected by Amazon SES.