Network.AWS.SES.VerifyDomainDkim

Description

Returns a set of DKIM tokens for a domain. DKIM tokens are character strings that represent your domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain. This action is throttled at one request per second. To enable or disable Easy DKIM signing for a domain, use the SetIdentityDkimEnabled action. For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide. POST / HTTP/1.1 Date: Fri, 29 Jun 2012 22:43:30 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=roXhd+JhEjeBBo5tSERhrptRHSw4XHz6Ra4BXyHIduk=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 136 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=VerifyDomainDkim &Domain=example.com &Timestamp=2012-06-29T22%3A43%3A30.000Z &Version=2010-12-01 vvjuipp74whm76gqoni7qmwwn4w4qusjiainivf6sf 3frqe7jn4obpuxjpwpolz6ipb3k5nvt2nhjpik2oy wrqplteh7oodxnad7hsl4mixg2uavzneazxv5sxi2 9662c15b-c469-11e1-99d1-797d6ecd6414.

Request

data VerifyDomainDkim

Represents a request instructing the service to begin DKIM verification for a domain.

Request constructor

verifyDomainDkim

Arguments

:: Text
vddDomain

-> VerifyDomainDkim

Smart constructor for the minimum required parameters to construct a valid VerifyDomainDkim request.

The fields accessible through corresponding lenses are:

Domain :: Text

Request lenses

vddDomain :: Lens' VerifyDomainDkim Text

The name of the domain to be verified for Easy DKIM signing.

Response

data VerifyDomainDkimResponse

Represents the DNS records that must be published in the domain name's DNS to complete DKIM setup.

Response constructor

verifyDomainDkimResponse

Arguments

:: [Text]
vddrDkimTokens

-> VerifyDomainDkimResponse

Smart constructor for the minimum required parameters to construct a valid VerifyDomainDkimResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DkimTokens :: [Text]

Response lenses

vddrDkimTokens :: Lens' VerifyDomainDkimResponse [Text]

A set of character strings that represent the domain's identity. If the identity is an email address, the tokens represent the domain of that address. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign emails originating from that domain. For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.