Network.AWS.SES.VerifyDomainIdentity

Description

Verifies a domain. This action is throttled at one request per second. POST HTTP1.1 Date: Sat, 12 May 2012 05:24:02 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=Wr+6RCfV+QgjLki2dtIrlecMK9+RrsDaTG5uWneDAu8=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 139 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=VerifyDomainIdentity &Domain=domain.com &Timestamp=2012-05-12T05%3A24%3A02.000Z &Version=2010-12-01 QTKknzFg2J4ygwa+XvHAxUl1hyHoY0gVfZdfjIedHZ0= 94f6368e-9bf2-11e1-8ee7-c98a0037a2b6.

Request

data VerifyDomainIdentity

Represents a request instructing the service to begin domain verification.

Request constructor

mkVerifyDomainIdentity

Arguments

:: Text
vdiDomain

-> VerifyDomainIdentity

Smart constructor for the minimum required parameters to construct a valid VerifyDomainIdentity request.

The fields accessible through corresponding lenses are:

Domain :: Text

Request lenses

vdiDomain :: Lens' VerifyDomainIdentity Text

The domain to be verified.

Response

data VerifyDomainIdentityResponse

Represents a token used for domain ownership verification.

Response constructor

mkVerifyDomainIdentityResponse

Arguments

:: Text
vdirVerificationToken

-> VerifyDomainIdentityResponse

Smart constructor for the minimum required parameters to construct a valid VerifyDomainIdentityResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VerificationToken :: Text

Response lenses

vdirVerificationToken :: Lens' VerifyDomainIdentityResponse Text

A TXT record that must be placed in the DNS settings for the domain, in order to complete domain verification.