Network.AWS.SES.GetIdentityDkimAttributes

Description

Returns the current status of Easy DKIM signing for an entity. For domain name identities, this action also returns the DKIM tokens that are required for Easy DKIM signing, and whether Amazon SES has successfully verified that these tokens have been published.

This action takes a list of identities as input and returns the following information for each:

Whether Easy DKIM signing is enabled or disabled.

A set of DKIM tokens that represent the identity. If the identity is an email address, the tokens represent the domain of that address.

Whether Amazon SES has successfully verified the DKIM tokens published in the domain's DNS. This information is only returned for domain name identities, not for email addresses.

This action is throttled at one request per second and can only get DKIM attributes for up to 100 identities at a time.

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide .

Creating a Request

getIdentityDkimAttributes :: GetIdentityDkimAttributes

Creates a value of GetIdentityDkimAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gidaIdentities - A list of one or more verified identities - email addresses, domains, or both.

data GetIdentityDkimAttributes

Represents a request for the status of Amazon SES Easy DKIM signing for an identity. For domain identities, this request also returns the DKIM tokens that are required for Easy DKIM signing, and whether Amazon SES successfully verified that these tokens were published. For more information about Easy DKIM, see the Amazon SES Developer Guide .

See: getIdentityDkimAttributes smart constructor.

Request Lenses

gidaIdentities :: Lens' GetIdentityDkimAttributes [Text]

A list of one or more verified identities - email addresses, domains, or both.

Destructuring the Response

getIdentityDkimAttributesResponse

Arguments

:: Int
gidarsResponseStatus

-> GetIdentityDkimAttributesResponse

Creates a value of GetIdentityDkimAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gidarsResponseStatus - -- | The response status code.

gidarsDkimAttributes - The DKIM attributes for an email address or a domain.

data GetIdentityDkimAttributesResponse

Represents the status of Amazon SES Easy DKIM signing for an identity. For domain identities, this response also contains the DKIM tokens that are required for Easy DKIM signing, and whether Amazon SES successfully verified that these tokens were published.

See: getIdentityDkimAttributesResponse smart constructor.

Response Lenses

gidarsResponseStatus :: Lens' GetIdentityDkimAttributesResponse Int

- | The response status code.

gidarsDkimAttributes :: Lens' GetIdentityDkimAttributesResponse (HashMap Text IdentityDkimAttributes)

The DKIM attributes for an email address or a domain.