Network.AWS.SES.GetIdentityDkimAttributes

Description

Returns the current status of Easy DKIM signing for an entity. For domain name identities, this action also returns the DKIM tokens that are required for Easy DKIM signing, and whether Amazon SES has successfully verified that these tokens have been published.

This action takes a list of identities as input and returns the following information for each:

Whether Easy DKIM signing is enabled or disabled. A set of DKIM tokens that represent the identity. If the identity is an email address, the tokens represent the domain of that address. Whether Amazon SES has successfully verified the DKIM tokens published in the domain's DNS. This information is only returned for domain name identities, not for email addresses. This action is throttled at one request per second.

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_GetIdentityDkimAttributes.html

Request

data GetIdentityDkimAttributes

Request constructor

getIdentityDkimAttributes :: GetIdentityDkimAttributes

GetIdentityDkimAttributes constructor.

The fields accessible through corresponding lenses are:

gidaIdentities :: [Text]

Request lenses

gidaIdentities :: Lens' GetIdentityDkimAttributes [Text]

A list of one or more verified identities - email addresses, domains, or both.

Response

data GetIdentityDkimAttributesResponse

Response constructor

getIdentityDkimAttributesResponse :: GetIdentityDkimAttributesResponse

GetIdentityDkimAttributesResponse constructor.

The fields accessible through corresponding lenses are:

gidarDkimAttributes :: HashMap Text IdentityDkimAttributes

Response lenses

gidarDkimAttributes :: Lens' GetIdentityDkimAttributesResponse (HashMap Text IdentityDkimAttributes)

The DKIM attributes for an email address or a domain.