Network.AWS.SES.GetIdentityNotificationAttributes

Description

Given a list of verified identities (email addresses and/or domains), returns a structure describing identity notification attributes.

This action is throttled at one request per second and can only get notification attributes for up to 100 identities at a time.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide .

Creating a Request

getIdentityNotificationAttributes :: GetIdentityNotificationAttributes

Creates a value of GetIdentityNotificationAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ginaIdentities - A list of one or more identities. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: userexample.com , example.com , arn:aws:ses:us-east-1:123456789012:identity/example.com@ .

data GetIdentityNotificationAttributes

Represents a request to return the notification attributes for a list of identities you verified with Amazon SES. For information about Amazon SES notifications, see the Amazon SES Developer Guide .

See: getIdentityNotificationAttributes smart constructor.

Request Lenses

ginaIdentities :: Lens' GetIdentityNotificationAttributes [Text]

A list of one or more identities. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: userexample.com , example.com , arn:aws:ses:us-east-1:123456789012:identity/example.com@ .

Destructuring the Response

getIdentityNotificationAttributesResponse

Arguments

:: Int
ginarsResponseStatus

-> GetIdentityNotificationAttributesResponse

Creates a value of GetIdentityNotificationAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ginarsResponseStatus - -- | The response status code.

ginarsNotificationAttributes - A map of Identity to IdentityNotificationAttributes.

data GetIdentityNotificationAttributesResponse

Represents the notification attributes for a list of identities.

See: getIdentityNotificationAttributesResponse smart constructor.

Response Lenses

ginarsResponseStatus :: Lens' GetIdentityNotificationAttributesResponse Int

- | The response status code.

ginarsNotificationAttributes :: Lens' GetIdentityNotificationAttributesResponse (HashMap Text IdentityNotificationAttributes)

A map of Identity to IdentityNotificationAttributes.