Network.AWS.SES.SendEmail

Description

Composes an email message based on input data, and then immediately queues the message for sending.

There are several important points to know about SendEmail :

You can only send email from verified email addresses and domains; otherwise, you will get an "Email address not verified" error. If your account is still in the Amazon SES sandbox, you must also verify every recipient email address except for the recipients provided by the Amazon SES mailbox simulator. For more information, go to the Amazon SES Developer Guide .

The total size of the message cannot exceed 10 MB. This includes any attachments that are part of the message.

You must provide at least one recipient email address. The recipient address can be a To: address, a CC: address, or a BCC: address. If any email address you provide is invalid, Amazon SES rejects the entire email.

Amazon SES has a limit on the total number of recipients per message. The combined number of To:, CC: and BCC: email addresses cannot exceed 50. If you need to send an email message to a larger audience, you can divide your recipient list into groups of 50 or fewer, and then call Amazon SES repeatedly to send the message to each group.

For every message that you send, the total number of recipients (To:, CC: and BCC:) is counted against your sending quota - the maximum number of emails you can send in a 24-hour period. For information about your sending quota, go to the Amazon SES Developer Guide .

Creating a Request

sendEmail

Arguments

:: Text
seSource

-> Destination
seDestination

-> Message
seMessage

-> SendEmail

Creates a value of SendEmail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

seReturnPath - The email address to which bounces and complaints are to be forwarded when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

seConfigurationSetName - The name of the configuration set to use when you send an email using SendEmail .

seSourceARN - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter. For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com ) attaches a policy to it that authorizes you to send from userexample.com , then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identityexample.com , and the Source to be userexample.com . For more information about sending authorization, see the <http:docs.aws.amazon.comseslatestDeveloperGuide/sending-authorization.html Amazon SES Developer Guide> .

seReturnPathARN - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter. For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com ) attaches a policy to it that authorizes you to use feedbackexample.com , then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identityexample.com , and the ReturnPath to be feedbackexample.com . For more information about sending authorization, see the <http:docs.aws.amazon.comseslatestDeveloperGuide/sending-authorization.html Amazon SES Developer Guide> .

seTags - A list of tags, in the form of name/value pairs, to apply to an email that you send using SendEmail . Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

seReplyToAddresses - The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

seSource - The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide . If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide . In all cases, the email address must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?= . For more information, see RFC 2047 .

seDestination - The destination for this email, composed of To:, CC:, and BCC: fields.

seMessage - The message to be sent.

data SendEmail

Represents a request to send a single formatted email using Amazon SES. For more information, see the Amazon SES Developer Guide .

See: sendEmail smart constructor.

Request Lenses

seReturnPath :: Lens' SendEmail (Maybe Text)

The email address to which bounces and complaints are to be forwarded when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

seConfigurationSetName :: Lens' SendEmail (Maybe Text)

The name of the configuration set to use when you send an email using SendEmail .

seSourceARN :: Lens' SendEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter. For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com ) attaches a policy to it that authorizes you to send from userexample.com , then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identityexample.com , and the Source to be userexample.com . For more information about sending authorization, see the <http:docs.aws.amazon.comseslatestDeveloperGuide/sending-authorization.html Amazon SES Developer Guide> .

seReturnPathARN :: Lens' SendEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter. For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com ) attaches a policy to it that authorizes you to use feedbackexample.com , then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identityexample.com , and the ReturnPath to be feedbackexample.com . For more information about sending authorization, see the <http:docs.aws.amazon.comseslatestDeveloperGuide/sending-authorization.html Amazon SES Developer Guide> .

seTags :: Lens' SendEmail [MessageTag]

A list of tags, in the form of name/value pairs, to apply to an email that you send using SendEmail . Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

seReplyToAddresses :: Lens' SendEmail [Text]

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

seSource :: Lens' SendEmail Text

The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide . If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide . In all cases, the email address must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?= . For more information, see RFC 2047 .

seDestination :: Lens' SendEmail Destination

The destination for this email, composed of To:, CC:, and BCC: fields.

seMessage :: Lens' SendEmail Message

The message to be sent.

Destructuring the Response

sendEmailResponse

Arguments

:: Int
sersResponseStatus

-> Text
sersMessageId

-> SendEmailResponse

Creates a value of SendEmailResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sersResponseStatus - -- | The response status code.

sersMessageId - The unique message identifier returned from the SendEmail action.

data SendEmailResponse

Represents a unique message ID.

See: sendEmailResponse smart constructor.

Response Lenses

sersResponseStatus :: Lens' SendEmailResponse Int

- | The response status code.

sersMessageId :: Lens' SendEmailResponse Text

The unique message identifier returned from the SendEmail action.