Network.AWS.SES.SendRawEmail

Description

Sends an email message, with header and content specified by the client. The SendRawEmail action is useful for sending multipart MIME emails. The raw text of the message must comply with Internet email standards; otherwise, the message cannot be sent.

You can only send email from verified email addresses and domains. If you have not requested production access to Amazon SES, you must also verify every recipient email address except for the recipients provided by the Amazon SES mailbox simulator. For more information, go to the Amazon SESDeveloper Guide. The total size of the message cannot exceed 10 MB. This includes any attachments that are part of the message.

Amazon SES has a limit on the total number of recipients per message: The combined number of To:, CC: and BCC: email addresses cannot exceed 50. If you need to send an email message to a larger audience, you can divide your recipient list into groups of 50 or fewer, and then call Amazon SES repeatedly to send the message to each group.

The To:, CC:, and BCC: headers in the raw message can contain a group list. Note that each recipient in a group list counts towards the 50-recipient limit.

For every message that you send, the total number of recipients (To:, CC: and BCC:) is counted against your sending quota - the maximum number of emails you can send in a 24-hour period. For information about your sending quota, go to the Amazon SES Developer Guide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_SendRawEmail.html

Request

data SendRawEmail

Request constructor

sendRawEmail

Arguments

:: RawMessage
sreRawMessage

-> SendRawEmail

SendRawEmail constructor.

The fields accessible through corresponding lenses are:

sreDestinations :: [Text]

sreRawMessage :: RawMessage

sreSource :: Maybe Text

Request lenses

sreDestinations :: Lens' SendRawEmail [Text]

A list of destinations for the message, consisting of To:, CC:, and BCC: addresses.

sreRawMessage :: Lens' SendRawEmail RawMessage

The raw text of the message. The client is responsible for ensuring the following:

Message must contain a header and a body, separated by a blank line. All required header fields must be present. Each part of a multipart MIME message must be formatted properly. MIME content types must be among those supported by Amazon SES. For more information, go to the Amazon SES Developer Guide. Content must be base64-encoded, if MIME requires it.

sreSource :: Lens' SendRawEmail (Maybe Text)

The identity's email address.

By default, the string must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: '=?charset?encoding?encoded-text?='. For more information, see RFC 2047.

Response

data SendRawEmailResponse

Response constructor

sendRawEmailResponse

Arguments

:: Text
srerMessageId

-> SendRawEmailResponse

SendRawEmailResponse constructor.

The fields accessible through corresponding lenses are:

srerMessageId :: Text

Response lenses

srerMessageId :: Lens' SendRawEmailResponse Text

The unique message identifier returned from the SendRawEmail action.