Network.AWS.SES.SetIdentityDkimEnabled

Description

Enables or disables Easy DKIM signing of email sent from an identity:

If Easy DKIM signing is enabled for a domain name identity (e.g., example.com ), then Amazon SES will DKIM-sign all email sent by addresses under that domain name (e.g., userexample.com@ ).

If Easy DKIM signing is enabled for an email address, then Amazon SES will DKIM-sign all email sent by that email address.

For email addresses (e.g., userexample.com ), you can only enable Easy DKIM signing if the corresponding domain (e.g., example.com ) has been set up for Easy DKIM using the AWS Console or the VerifyDomainDkim@ action.

This action is throttled at one request per second.

For more information about Easy DKIM signing, go to the Amazon SES Developer Guide .

Creating a Request

setIdentityDkimEnabled

Arguments

:: Text
sideIdentity

-> Bool
sideDkimEnabled

-> SetIdentityDkimEnabled

Creates a value of SetIdentityDkimEnabled with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sideIdentity - The identity for which DKIM signing should be enabled or disabled.

sideDkimEnabled - Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

data SetIdentityDkimEnabled

Represents a request to enable or disable Amazon SES Easy DKIM signing for an identity. For more information about setting up Easy DKIM, see the Amazon SES Developer Guide .

See: setIdentityDkimEnabled smart constructor.

Request Lenses

sideIdentity :: Lens' SetIdentityDkimEnabled Text

The identity for which DKIM signing should be enabled or disabled.

sideDkimEnabled :: Lens' SetIdentityDkimEnabled Bool

Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

Destructuring the Response

setIdentityDkimEnabledResponse

Arguments

:: Int
sidersResponseStatus

-> SetIdentityDkimEnabledResponse

Creates a value of SetIdentityDkimEnabledResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sidersResponseStatus - -- | The response status code.

data SetIdentityDkimEnabledResponse

An empty element returned on a successful request.

See: setIdentityDkimEnabledResponse smart constructor.

Response Lenses

sidersResponseStatus :: Lens' SetIdentityDkimEnabledResponse Int

- | The response status code.