Network.AWS.SES.SetIdentityFeedbackForwardingEnabled

Description

Given an identity (email address or domain), enables or disables whether Amazon SES forwards bounce and complaint notifications as email. Feedback forwarding can only be disabled when Amazon Simple Notification Service (Amazon SNS) topics are specified for both bounces and complaints.

This action is throttled at one request per second.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_SetIdentityFeedbackForwardingEnabled.html

Request

data SetIdentityFeedbackForwardingEnabled

Request constructor

setIdentityFeedbackForwardingEnabled

Arguments

:: Text
siffeIdentity

-> Bool
siffeForwardingEnabled

-> SetIdentityFeedbackForwardingEnabled

SetIdentityFeedbackForwardingEnabled constructor.

The fields accessible through corresponding lenses are:

siffeForwardingEnabled :: Bool

siffeIdentity :: Text

Request lenses

siffeForwardingEnabled :: Lens' SetIdentityFeedbackForwardingEnabled Bool

Sets whether Amazon SES will forward bounce and complaint notifications as email. true specifies that Amazon SES will forward bounce and complaint notifications as email, in addition to any Amazon SNS topic publishing otherwise specified. false specifies that Amazon SES will publish bounce and complaint notifications only through Amazon SNS. This value can only be set to false when Amazon SNS topics are set for both Bounce and Complaint notification types.

siffeIdentity :: Lens' SetIdentityFeedbackForwardingEnabled Text

The identity for which to set bounce and complaint notification forwarding. Examples: 'user@example.com', 'example.com'.

Response

data SetIdentityFeedbackForwardingEnabledResponse

Response constructor

setIdentityFeedbackForwardingEnabledResponse :: SetIdentityFeedbackForwardingEnabledResponse

SetIdentityFeedbackForwardingEnabledResponse constructor.