Network.AWS.SES.Types

Service

data SES

Version 2010-12-01 of the Amazon Simple Email Service service.

Error

data RESTError :: *

XML

ns :: Text

Destination

data Destination

destination :: Destination

Destination constructor.

The fields accessible through corresponding lenses are:

dBccAddresses :: [Text]

dCcAddresses :: [Text]

dToAddresses :: [Text]

dBccAddresses :: Lens' Destination [Text]

The BCC: field(s) of the message.

dCcAddresses :: Lens' Destination [Text]

The CC: field(s) of the message.

dToAddresses :: Lens' Destination [Text]

The To: field(s) of the message.

IdentityDkimAttributes

data IdentityDkimAttributes

identityDkimAttributes

Arguments

:: Bool
idaDkimEnabled

-> VerificationStatus
idaDkimVerificationStatus

-> IdentityDkimAttributes

IdentityDkimAttributes constructor.

The fields accessible through corresponding lenses are:

idaDkimEnabled :: Bool

idaDkimTokens :: [Text]

idaDkimVerificationStatus :: VerificationStatus

idaDkimEnabled :: Lens' IdentityDkimAttributes Bool

True if DKIM signing is enabled for email sent from the identity; false otherwise.

idaDkimTokens :: Lens' IdentityDkimAttributes [Text]

A set of character strings that represent the domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain. (This only applies to domain identities, not email address identities.)

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

idaDkimVerificationStatus :: Lens' IdentityDkimAttributes VerificationStatus

Describes whether Amazon SES has successfully verified the DKIM DNS records (tokens) published in the domain name's DNS. (This only applies to domain identities, not email address identities.)

Body

data Body

body :: Body

Body constructor.

The fields accessible through corresponding lenses are:

bHtml :: Maybe Content

bText :: Maybe Content

bHtml :: Lens' Body (Maybe Content)

The content of the message, in HTML format. Use this for email clients that can process HTML. You can include clickable links, formatted text, and much more in an HTML message.

bText :: Lens' Body (Maybe Content)

The content of the message, in text format. Use this for text-based email clients, or clients on high-latency networks (such as mobile devices).

IdentityVerificationAttributes

data IdentityVerificationAttributes

identityVerificationAttributes

Arguments

:: VerificationStatus
ivaVerificationStatus

-> IdentityVerificationAttributes

IdentityVerificationAttributes constructor.

The fields accessible through corresponding lenses are:

ivaVerificationStatus :: VerificationStatus

ivaVerificationToken :: Maybe Text

ivaVerificationStatus :: Lens' IdentityVerificationAttributes VerificationStatus

The verification status of the identity: Pending, Success, Failed, or TemporaryFailure.

ivaVerificationToken :: Lens' IdentityVerificationAttributes (Maybe Text)

The verification token for a domain identity. Null for email address identities.

SendDataPoint

data SendDataPoint

sendDataPoint :: SendDataPoint

SendDataPoint constructor.

The fields accessible through corresponding lenses are:

sdpBounces :: Maybe Integer

sdpComplaints :: Maybe Integer

sdpDeliveryAttempts :: Maybe Integer

sdpRejects :: Maybe Integer

sdpTimestamp :: Maybe UTCTime

sdpBounces :: Lens' SendDataPoint (Maybe Integer)

Number of emails that have bounced.

sdpComplaints :: Lens' SendDataPoint (Maybe Integer)

Number of unwanted emails that were rejected by recipients.

sdpDeliveryAttempts :: Lens' SendDataPoint (Maybe Integer)

Number of emails that have been enqueued for sending.

sdpRejects :: Lens' SendDataPoint (Maybe Integer)

Number of emails rejected by Amazon SES.

sdpTimestamp :: Lens' SendDataPoint (Maybe UTCTime)

Time of the data point.

IdentityType

data IdentityType

Constructors

ITDomain
Domain

ITEmailAddress
EmailAddress

Content

data Content

content

Arguments

:: Text
cData

-> Content

Content constructor.

The fields accessible through corresponding lenses are:

cCharset :: Maybe Text

cData :: Text

cCharset :: Lens' Content (Maybe Text)

The character set of the content.

cData :: Lens' Content Text

The textual data of the content.

IdentityNotificationAttributes

data IdentityNotificationAttributes

identityNotificationAttributes

Arguments

:: Text
inaBounceTopic

-> Text
inaComplaintTopic

-> Text
inaDeliveryTopic

-> Bool
inaForwardingEnabled

-> IdentityNotificationAttributes

IdentityNotificationAttributes constructor.

The fields accessible through corresponding lenses are:

inaBounceTopic :: Text

inaComplaintTopic :: Text

inaDeliveryTopic :: Text

inaForwardingEnabled :: Bool

inaBounceTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish bounce notifications.

inaComplaintTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish complaint notifications.

inaDeliveryTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish delivery notifications.

inaForwardingEnabled :: Lens' IdentityNotificationAttributes Bool

Describes whether Amazon SES will forward bounce and complaint notifications as email. true indicates that Amazon SES will forward bounce and complaint notifications as email, while false indicates that bounce and complaint notifications will be published only to the specified bounce and complaint Amazon SNS topics.

RawMessage

data RawMessage

rawMessage

Arguments

:: Base64
rmData

-> RawMessage

RawMessage constructor.

The fields accessible through corresponding lenses are:

rmData :: Base64

rmData :: Lens' RawMessage Base64

The raw data of the message. The client must ensure that the message format complies with Internet email standards regarding email header fields, MIME types, MIME encoding, and base64 encoding (if necessary).

The To:, CC:, and BCC: headers in the raw message can contain a group list.

For more information, go to the Amazon SES Developer Guide.

NotificationType

data NotificationType

Constructors

Bounce
Bounce

Complaint
Complaint

Delivery
Delivery

VerificationStatus

data VerificationStatus

Constructors

Failed
Failed

NotStarted
NotStarted

Pending
Pending

Success
Success

TemporaryFailure
TemporaryFailure

Message

data Message

message

Arguments

:: Content
mSubject

-> Body
mBody

-> Message

Message constructor.

The fields accessible through corresponding lenses are:

mBody :: Body

mSubject :: Content

mBody :: Lens' Message Body

The message body.

mSubject :: Lens' Message Content

The subject of the message: A short summary of the content, which will appear in the recipient's inbox.