Network.AWS.SES.V2010_12_01

Description

Amazon Simple Email Service (Amazon SES) is a cost-effective outbound-only email-sending service built on the reliable and scalable infrastructure that Amazon.com has developed to serve its own customer base. With Amazon SES, you can send transactional email, marketing messages, or any other type of high-quality content and you only pay for what you use. Along with high deliverability, Amazon SES provides easy, real-time access to your sending statistics and built-in notifications for bounces, complaints, and deliveries to help you fine-tune your email-sending strategy.

Documentation

data DeleteIdentity

Constructors

DeleteIdentity

Fields

_dirIdentity :: Text

The identity to be removed from the list of identities for the AWS Account.

dirIdentity :: Iso' DeleteIdentity Text

data DeleteIdentityResponse

Constructors

DeleteIdentityResponse

data DeleteVerifiedEmailAddress

Constructors

DeleteVerifiedEmailAddress

Fields

_dvearEmailAddress :: Text

An email address to be removed from the list of verified addresses.

dvearEmailAddress :: Iso' DeleteVerifiedEmailAddress Text

data DeleteVerifiedEmailAddressResponse

Constructors

DeleteVerifiedEmailAddressResponse

data GetIdentityDkimAttributes

Constructors

GetIdentityDkimAttributes

Fields

_gidarIdentities :: [Text]

A list of one or more verified identities - email addresses, domains, or both.

gidarIdentities :: Iso' GetIdentityDkimAttributes [Text]

data GetIdentityDkimAttributesResponse

Constructors

GetIdentityDkimAttributesResponse

Fields

_gidasDkimAttributes :: Map Text IdentityDkimAttributes

The DKIM attributes for an email address or a domain.

gidasDkimAttributes :: Iso' GetIdentityDkimAttributesResponse (Map Text IdentityDkimAttributes)

data GetIdentityNotificationAttributes

Constructors

GetIdentityNotificationAttributes

Fields

_ginarIdentities :: [Text]

A list of one or more identities.

ginarIdentities :: Iso' GetIdentityNotificationAttributes [Text]

data GetIdentityNotificationAttributesResponse

Constructors

GetIdentityNotificationAttributesResponse

Fields

_ginasNotificationAttributes :: Map Text IdentityNotificationAttributes

A map of Identity to IdentityNotificationAttributes.

ginasNotificationAttributes :: Iso' GetIdentityNotificationAttributesResponse (Map Text IdentityNotificationAttributes)

data GetIdentityVerificationAttributes

Constructors

GetIdentityVerificationAttributes

Fields

_givarIdentities :: [Text]

A list of identities.

givarIdentities :: Iso' GetIdentityVerificationAttributes [Text]

data GetIdentityVerificationAttributesResponse

Constructors

GetIdentityVerificationAttributesResponse

Fields

_givasVerificationAttributes :: Map Text IdentityVerificationAttributes

A map of Identities to IdentityVerificationAttributes objects.

givasVerificationAttributes :: Iso' GetIdentityVerificationAttributesResponse (Map Text IdentityVerificationAttributes)

data GetSendQuota

Constructors

GetSendQuota

data GetSendQuotaResponse

Constructors

GetSendQuotaResponse

Fields

_gsqrMax24HourSend :: Maybe Double

The maximum number of emails the user is allowed to send in a 24-hour interval.

_gsqrMaxSendRate :: Maybe Double

The maximum number of emails the user is allowed to send per second.

_gsqrSentLast24Hours :: Maybe Double

The number of emails sent during the previous 24 hours.

gsqrSentLast24Hours :: Lens' GetSendQuotaResponse (Maybe Double)

gsqrMaxSendRate :: Lens' GetSendQuotaResponse (Maybe Double)

gsqrMax24HourSend :: Lens' GetSendQuotaResponse (Maybe Double)

data GetSendStatistics

Constructors

GetSendStatistics

data GetSendStatisticsResponse

Constructors

GetSendStatisticsResponse

Fields

_gssrSendDataPoints :: [SendDataPoint]

A list of data points, each of which represents 15 minutes of activity.

gssrSendDataPoints :: Iso' GetSendStatisticsResponse [SendDataPoint]

listIdentities :: ListIdentities

Minimum specification for a ListIdentities request.

data ListIdentities

Constructors

ListIdentities

Fields

_lirIdentityType :: Maybe IdentityType

The type of the identities to list. Possible values are EmailAddress and Domain. If this parameter is omitted, then all identities will be listed.

_lirMaxItems :: Maybe Integer

The maximum number of identities per page. Possible values are 1-100 inclusive.

_lirNextToken :: Maybe Text

The token to use for pagination.

lirNextToken :: Lens' ListIdentities (Maybe Text)

lirMaxItems :: Lens' ListIdentities (Maybe Integer)

lirIdentityType :: Lens' ListIdentities (Maybe IdentityType)

data ListIdentitiesResponse

Constructors

ListIdentitiesResponse

Fields

_lisIdentities :: [Text]

A list of identities.

_lisNextToken :: Maybe Text

The token used for pagination.

lisNextToken :: Lens' ListIdentitiesResponse (Maybe Text)

lisIdentities :: Lens' ListIdentitiesResponse [Text]

data ListVerifiedEmailAddresses

Constructors

ListVerifiedEmailAddresses

data ListVerifiedEmailAddressesResponse

Constructors

ListVerifiedEmailAddressesResponse

Fields

_lvearVerifiedEmailAddresses :: [Text]

A list of email addresses that have been verified.

lvearVerifiedEmailAddresses :: Iso' ListVerifiedEmailAddressesResponse [Text]

sendEmail

Arguments

:: Text
_serSource

-> Destination
_serDestination

-> Message
_serMessage

-> SendEmail

Minimum specification for a SendEmail request.

data SendEmail

Constructors

SendEmail

Fields

_serSource :: Text

The identity's email address. By default, the string must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=. For more information, see RFC 2047.

_serDestination :: Destination

The destination for this email, composed of To:, CC:, and BCC: fields.

_serMessage :: Message

The message to be sent.

_serReturnPath :: Maybe Text

The email address to which bounces and complaints are to be forwarded when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter.

_serReplyToAddresses :: [Text]

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

serSource :: Lens' SendEmail Text

serReturnPath :: Lens' SendEmail (Maybe Text)

serReplyToAddresses :: Lens' SendEmail [Text]

serMessage :: Lens' SendEmail Message

serDestination :: Lens' SendEmail Destination

data SendEmailResponse

Constructors

SendEmailResponse

Fields

_sesMessageId :: Text

The unique message identifier returned from the SendEmail action.

sesMessageId :: Iso' SendEmailResponse Text

sendRawEmail

Arguments

:: RawMessage
_srerRawMessage

-> SendRawEmail

Minimum specification for a SendRawEmail request.

data SendRawEmail

Constructors

SendRawEmail

Fields

_srerRawMessage :: RawMessage

The raw text of the message. The client is responsible for ensuring the following: Message must contain a header and a body, separated by a blank line. All required header fields must be present. Each part of a multipart MIME message must be formatted properly. MIME content types must be among those supported by Amazon SES. For more information, go to the Amazon SES Developer Guide. Content must be base64-encoded, if MIME requires it.

_srerSource :: Maybe Text

The identity's email address. By default, the string must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=. For more information, see RFC 2047. If you specify the Source parameter and have feedback forwarding enabled, then bounces and complaints will be sent to this email address. This takes precedence over any Return-Path header that you might include in the raw text of the message.

_srerDestinations :: [Text]

A list of destinations for the message, consisting of To:, CC:, and BCC: addresses.

srerSource :: Lens' SendRawEmail (Maybe Text)

srerRawMessage :: Lens' SendRawEmail RawMessage

srerDestinations :: Lens' SendRawEmail [Text]

data SendRawEmailResponse

Constructors

SendRawEmailResponse

Fields

_sresMessageId :: Text

The unique message identifier returned from the SendRawEmail action.

sresMessageId :: Iso' SendRawEmailResponse Text

data SetIdentityDkimEnabled

Constructors

SetIdentityDkimEnabled

Fields

_siderDkimEnabled :: Bool

Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

_siderIdentity :: Text

The identity for which DKIM signing should be enabled or disabled.

siderIdentity :: Lens' SetIdentityDkimEnabled Text

siderDkimEnabled :: Lens' SetIdentityDkimEnabled Bool

data SetIdentityDkimEnabledResponse

Constructors

SetIdentityDkimEnabledResponse

data SetIdentityFeedbackForwardingEnabled

Constructors

SetIdentityFeedbackForwardingEnabled

Fields

_sifferForwardingEnabled :: Bool

Sets whether Amazon SES will forward bounce and complaint notifications as email. true specifies that Amazon SES will forward bounce and complaint notifications as email, in addition to any Amazon SNS topic publishing otherwise specified. false specifies that Amazon SES will publish bounce and complaint notifications only through Amazon SNS. This value can only be set to false when Amazon SNS topics are set for both Bounce and Complaint notification types.

_sifferIdentity :: Text

The identity for which to set bounce and complaint notification forwarding. Examples: user@example.com, example.com.

sifferIdentity :: Lens' SetIdentityFeedbackForwardingEnabled Text

sifferForwardingEnabled :: Lens' SetIdentityFeedbackForwardingEnabled Bool

data SetIdentityFeedbackForwardingEnabledResponse

Constructors

SetIdentityFeedbackForwardingEnabledResponse

setIdentityNotificationTopic

Arguments

:: Text
_sintrIdentity

-> NotificationType
_sintrNotificationType

-> SetIdentityNotificationTopic

Minimum specification for a SetIdentityNotificationTopic request.

data SetIdentityNotificationTopic

Constructors

SetIdentityNotificationTopic

Fields

_sintrIdentity :: Text

The identity for which the Amazon SNS topic will be set. Examples: user@example.com, example.com.

_sintrNotificationType :: NotificationType

The type of notifications that will be published to the specified Amazon SNS topic.

_sintrSnsTopic :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SNS topic. If the parameter is omitted from the request or a null value is passed, SnsTopic is cleared and publishing is disabled.

sintrSnsTopic :: Lens' SetIdentityNotificationTopic (Maybe Text)

sintrNotificationType :: Lens' SetIdentityNotificationTopic NotificationType

sintrIdentity :: Lens' SetIdentityNotificationTopic Text

data SetIdentityNotificationTopicResponse

Constructors

SetIdentityNotificationTopicResponse

data SES

Supported version (2010-12-01) of the Amazon Simple Email Service service.

xmlOptions :: Tagged a XMLOptions

data IdentityType

The type of the identities to list. Possible values are EmailAddress and Domain. If this parameter is omitted, then all identities will be listed.

Constructors

IdentityTypeDomain
Domain

IdentityTypeEmailAddress
EmailAddress

data NotificationType

The type of notifications that will be published to the specified Amazon SNS topic.

Constructors

NotificationTypeBounce
Bounce

NotificationTypeComplaint
Complaint

NotificationTypeDelivery
Delivery

data VerificationStatus

The verification status of the identity: Pending, Success, Failed, or TemporaryFailure.

Constructors

VerificationStatusFailed
Failed

VerificationStatusNotStarted
NotStarted

VerificationStatusPending
Pending

VerificationStatusSuccess
Success

VerificationStatusTemporaryFailure
TemporaryFailure

newtype RawMessage

The raw text of the message. The client is responsible for ensuring the following: Message must contain a header and a body, separated by a blank line. All required header fields must be present. Each part of a multipart MIME message must be formatted properly. MIME content types must be among those supported by Amazon SES. For more information, go to the Amazon SES Developer Guide. Content must be base64-encoded, if MIME requires it.

Constructors

RawMessage

Fields

_rmData :: ByteString

The raw data of the message. The client must ensure that the message format complies with Internet email standards regarding email header fields, MIME types, MIME encoding, and base64 encoding (if necessary). The To:, CC:, and BCC: headers in the raw message can contain a group list. For more information, go to the Amazon SES Developer Guide.

data Body

The message body.

Constructors

Body

Fields

_byText :: Maybe Content

The content of the message, in text format. Use this for text-based email clients, or clients on high-latency networks (such as mobile devices).

_byHtml :: Maybe Content

The content of the message, in HTML format. Use this for email clients that can process HTML. You can include clickable links, formatted text, and much more in an HTML message.

data Content

The subject of the message: A short summary of the content, which will appear in the recipient's inbox.

Constructors

Content

Fields

_ctData :: Text

The textual data of the content.

_ctCharset :: Maybe Text

The character set of the content.

data Destination

The destination for this email, composed of To:, CC:, and BCC: fields.

Constructors

Destination

Fields

_dnBccAddresses :: [Text]

The BCC: field(s) of the message.

_dnCcAddresses :: [Text]

The CC: field(s) of the message.

_dnToAddresses :: [Text]

The To: field(s) of the message.

data IdentityDkimAttributes

Represents the DKIM attributes of a verified email address or a domain.

Constructors

IdentityDkimAttributes

Fields

_idaDkimTokens :: [Text]

A set of character strings that represent the domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain. (This only applies to domain identities, not email address identities.) For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

_idaDkimEnabled :: Bool

True if DKIM signing is enabled for email sent from the identity; false otherwise.

_idaDkimVerificationStatus :: VerificationStatus

Describes whether Amazon SES has successfully verified the DKIM DNS records (tokens) published in the domain name's DNS. (This only applies to domain identities, not email address identities.).

data IdentityNotificationAttributes

Represents the notification attributes of an identity, including whether an identity has Amazon Simple Notification Service (Amazon SNS) topics set for bounce, complaint, and/or delivery notifications, and whether feedback forwarding is enabled for bounce and complaint notifications.

Constructors

IdentityNotificationAttributes

Fields

_inaDeliveryTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish delivery notifications.

_inaBounceTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish bounce notifications.

_inaComplaintTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish complaint notifications.

_inaForwardingEnabled :: Bool

Describes whether Amazon SES will forward bounce and complaint notifications as email. true indicates that Amazon SES will forward bounce and complaint notifications as email, while false indicates that bounce and complaint notifications will be published only to the specified bounce and complaint Amazon SNS topics.

data IdentityVerificationAttributes

Represents the verification attributes of a single identity.

Constructors

IdentityVerificationAttributes

Fields

_ivaVerificationToken :: Maybe Text

The verification token for a domain identity. Null for email address identities.

_ivaVerificationStatus :: VerificationStatus

The verification status of the identity: Pending, Success, Failed, or TemporaryFailure.

data Message

The message to be sent.

Constructors

Message

Fields

_meSubject :: Content

The subject of the message: A short summary of the content, which will appear in the recipient's inbox.

_meBody :: Body

The message body.

data SendDataPoint

Represents sending statistics data. Each SendDataPoint contains statistics for a 15-minute period of sending activity.

Constructors

SendDataPoint

Fields

_sdpRejects :: Maybe Integer

Number of emails rejected by Amazon SES.

_sdpComplaints :: Maybe Integer

Number of unwanted emails that were rejected by recipients.

_sdpDeliveryAttempts :: Maybe Integer

Number of emails that have been enqueued for sending.

_sdpBounces :: Maybe Integer

Number of emails that have bounced.

_sdpTimestamp :: Maybe ISO8601

Time of the data point.

rmData :: Iso' RawMessage ByteString

byText :: Lens' Body (Maybe Content)

byHtml :: Lens' Body (Maybe Content)

ctData :: Lens' Content Text

ctCharset :: Lens' Content (Maybe Text)

dnToAddresses :: Lens' Destination [Text]

dnCcAddresses :: Lens' Destination [Text]

dnBccAddresses :: Lens' Destination [Text]

idaDkimVerificationStatus :: Lens' IdentityDkimAttributes VerificationStatus

idaDkimTokens :: Lens' IdentityDkimAttributes [Text]

idaDkimEnabled :: Lens' IdentityDkimAttributes Bool

inaForwardingEnabled :: Lens' IdentityNotificationAttributes Bool

inaDeliveryTopic :: Lens' IdentityNotificationAttributes Text

inaComplaintTopic :: Lens' IdentityNotificationAttributes Text

inaBounceTopic :: Lens' IdentityNotificationAttributes Text

ivaVerificationToken :: Lens' IdentityVerificationAttributes (Maybe Text)

ivaVerificationStatus :: Lens' IdentityVerificationAttributes VerificationStatus

meSubject :: Lens' Message Content

meBody :: Lens' Message Body

sdpTimestamp :: Lens' SendDataPoint (Maybe ISO8601)

sdpRejects :: Lens' SendDataPoint (Maybe Integer)

sdpDeliveryAttempts :: Lens' SendDataPoint (Maybe Integer)

sdpComplaints :: Lens' SendDataPoint (Maybe Integer)

sdpBounces :: Lens' SendDataPoint (Maybe Integer)

data VerifyDomainDkim

Constructors

VerifyDomainDkim

Fields

_vddrDomain :: Text

The name of the domain to be verified for Easy DKIM signing.

vddrDomain :: Iso' VerifyDomainDkim Text

data VerifyDomainDkimResponse

Constructors

VerifyDomainDkimResponse

Fields

_vddsDkimTokens :: [Text]

A set of character strings that represent the domain's identity. If the identity is an email address, the tokens represent the domain of that address. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign emails originating from that domain. For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

vddsDkimTokens :: Iso' VerifyDomainDkimResponse [Text]

data VerifyDomainIdentity

Constructors

VerifyDomainIdentity

Fields

_vdirDomain :: Text

The domain to be verified.

vdirDomain :: Iso' VerifyDomainIdentity Text

data VerifyDomainIdentityResponse

Constructors

VerifyDomainIdentityResponse

Fields

_vdisVerificationToken :: Text

A TXT record that must be placed in the DNS settings for the domain, in order to complete domain verification.

vdisVerificationToken :: Iso' VerifyDomainIdentityResponse Text

data VerifyEmailAddress

Constructors

VerifyEmailAddress

Fields

_vearEmailAddress :: Text

The email address to be verified.

vearEmailAddress :: Iso' VerifyEmailAddress Text

data VerifyEmailAddressResponse

Constructors

VerifyEmailAddressResponse

data VerifyEmailIdentity

Constructors

VerifyEmailIdentity

Fields

_veirEmailAddress :: Text

The email address to be verified.

veirEmailAddress :: Iso' VerifyEmailIdentity Text

data VerifyEmailIdentityResponse

Constructors

VerifyEmailIdentityResponse