Network.AWS.SES.VerifyDomainDkim

Description

Returns a set of DKIM tokens for a domain. DKIM tokens are character strings that represent your domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain.

This action is throttled at one request per second.

To enable or disable Easy DKIM signing for a domain, use the SetIdentityDkimEnabled action.

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide .

Creating a Request

verifyDomainDkim

Arguments

:: Text
vddDomain

-> VerifyDomainDkim

Creates a value of VerifyDomainDkim with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vddDomain - The name of the domain to be verified for Easy DKIM signing.

data VerifyDomainDkim

Represents a request to generate the CNAME records needed to set up Easy DKIM with Amazon SES. For more information about setting up Easy DKIM, see the Amazon SES Developer Guide .

See: verifyDomainDkim smart constructor.

Request Lenses

vddDomain :: Lens' VerifyDomainDkim Text

The name of the domain to be verified for Easy DKIM signing.

Destructuring the Response

verifyDomainDkimResponse

Arguments

:: Int
vddrsResponseStatus

-> VerifyDomainDkimResponse

Creates a value of VerifyDomainDkimResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vddrsResponseStatus - -- | The response status code.

vddrsDkimTokens - A set of character strings that represent the domain's identity. If the identity is an email address, the tokens represent the domain of that address. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign emails originating from that domain. For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide .

data VerifyDomainDkimResponse

Returns CNAME records that you must publish to the DNS server of your domain to set up Easy DKIM with Amazon SES.

See: verifyDomainDkimResponse smart constructor.

Response Lenses

vddrsResponseStatus :: Lens' VerifyDomainDkimResponse Int

- | The response status code.

vddrsDkimTokens :: Lens' VerifyDomainDkimResponse [Text]

A set of character strings that represent the domain's identity. If the identity is an email address, the tokens represent the domain of that address. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign emails originating from that domain. For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide .