Network.AWS.SES.VerifyDomainDkim

Description

Returns a set of DKIM tokens for a domain. DKIM tokens are character strings that represent your domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain.

This action is throttled at one request per second.

To enable or disable Easy DKIM signing for a domain, use the SetIdentityDkimEnabled action.

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

http://docs.aws.amazon.com/ses/latest/APIReference/API_VerifyDomainDkim.html

Request

data VerifyDomainDkim

Request constructor

verifyDomainDkim

Arguments

:: Text
vddDomain

-> VerifyDomainDkim

VerifyDomainDkim constructor.

The fields accessible through corresponding lenses are:

vddDomain :: Text

Request lenses

vddDomain :: Lens' VerifyDomainDkim Text

The name of the domain to be verified for Easy DKIM signing.

Response

data VerifyDomainDkimResponse

Response constructor

verifyDomainDkimResponse :: VerifyDomainDkimResponse

VerifyDomainDkimResponse constructor.

The fields accessible through corresponding lenses are:

vddrDkimTokens :: [Text]

Response lenses

vddrDkimTokens :: Lens' VerifyDomainDkimResponse [Text]

A set of character strings that represent the domain's identity. If the identity is an email address, the tokens represent the domain of that address.

Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign emails originating from that domain.

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.