Network.AWS.SES.VerifyDomainIdentity

Description

Verifies a domain.

This action is throttled at one request per second.

Creating a Request

verifyDomainIdentity

Arguments

:: Text
vdiDomain

-> VerifyDomainIdentity

Creates a value of VerifyDomainIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vdiDomain - The domain to be verified.

data VerifyDomainIdentity

Represents a request to begin Amazon SES domain verification and to generate the TXT records that you must publish to the DNS server of your domain to complete the verification. For information about domain verification, see the Amazon SES Developer Guide .

See: verifyDomainIdentity smart constructor.

Request Lenses

vdiDomain :: Lens' VerifyDomainIdentity Text

The domain to be verified.

Destructuring the Response

verifyDomainIdentityResponse

Arguments

:: Int
vdirsResponseStatus

-> Text
vdirsVerificationToken

-> VerifyDomainIdentityResponse

Creates a value of VerifyDomainIdentityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vdirsResponseStatus - -- | The response status code.

vdirsVerificationToken - A TXT record that you must place in the DNS settings of the domain to complete domain verification with Amazon SES. As Amazon SES searches for the TXT record, the domain's verification status is Pending. When Amazon SES detects the record, the domain's verification status changes to Success. If Amazon SES is unable to detect the record within 72 hours, the domain's verification status changes to "Failed." In that case, if you still want to verify the domain, you must restart the verification process from the beginning.

data VerifyDomainIdentityResponse

Returns a TXT record that you must publish to the DNS server of your domain to complete domain verification with Amazon SES.

See: verifyDomainIdentityResponse smart constructor.

Response Lenses

vdirsResponseStatus :: Lens' VerifyDomainIdentityResponse Int

- | The response status code.

vdirsVerificationToken :: Lens' VerifyDomainIdentityResponse Text

A TXT record that you must place in the DNS settings of the domain to complete domain verification with Amazon SES. As Amazon SES searches for the TXT record, the domain's verification status is Pending. When Amazon SES detects the record, the domain's verification status changes to Success. If Amazon SES is unable to detect the record within 72 hours, the domain's verification status changes to "Failed." In that case, if you still want to verify the domain, you must restart the verification process from the beginning.