Network.AWS.SES.VerifyDomainIdentity

Description

Verifies a domain.

This action is throttled at one request per second.

http://docs.aws.amazon.com/ses/latest/APIReference/API_VerifyDomainIdentity.html

Request

data VerifyDomainIdentity

Request constructor

verifyDomainIdentity

Arguments

:: Text
vdiDomain

-> VerifyDomainIdentity

VerifyDomainIdentity constructor.

The fields accessible through corresponding lenses are:

vdiDomain :: Text

Request lenses

vdiDomain :: Lens' VerifyDomainIdentity Text

The domain to be verified.

Response

data VerifyDomainIdentityResponse

Response constructor

verifyDomainIdentityResponse

Arguments

:: Text
vdirVerificationToken

-> VerifyDomainIdentityResponse

VerifyDomainIdentityResponse constructor.

The fields accessible through corresponding lenses are:

vdirVerificationToken :: Text

Response lenses

vdirVerificationToken :: Lens' VerifyDomainIdentityResponse Text

A TXT record that must be placed in the DNS settings for the domain, in order to complete domain verification.