Network.AWS.SES.VerifyEmailIdentity

Description

Verifies an email address. This action causes a confirmation email message to be sent to the specified address.

This action is throttled at one request per second.

Creating a Request

verifyEmailIdentity

Arguments

:: Text
veiEmailAddress

-> VerifyEmailIdentity

Creates a value of VerifyEmailIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

veiEmailAddress - The email address to be verified.

data VerifyEmailIdentity

Represents a request to begin email address verification with Amazon SES. For information about email address verification, see the Amazon SES Developer Guide .

See: verifyEmailIdentity smart constructor.

Request Lenses

veiEmailAddress :: Lens' VerifyEmailIdentity Text

The email address to be verified.

Destructuring the Response

verifyEmailIdentityResponse

Arguments

:: Int
veirsResponseStatus

-> VerifyEmailIdentityResponse

Creates a value of VerifyEmailIdentityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

veirsResponseStatus - -- | The response status code.

data VerifyEmailIdentityResponse

An empty element returned on a successful request.

See: verifyEmailIdentityResponse smart constructor.

Response Lenses

veirsResponseStatus :: Lens' VerifyEmailIdentityResponse Int

- | The response status code.